% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL.R
\name{adjust_PROSPECT_2_SAIL}
\alias{adjust_PROSPECT_2_SAIL}
\title{adjust prospect inputs and run prospect in preparation for 4SAIL or 4SAIL2}
\usage{
adjust_PROSPECT_2_SAIL(
  SAILversion,
  Spec_Sensor,
  Input_PROSPECT,
  CHL,
  CAR,
  ANT,
  BROWN,
  EWT,
  LMA,
  PROT,
  CBC,
  N,
  alpha,
  fraction_brown,
  BrownLOP = NULL
)
}
\arguments{
\item{SAILversion}{character. should be '4SAIL' or '4SAIL2'}

\item{Spec_Sensor}{dataframe. spectral properties required to run PROSPECT}

\item{Input_PROSPECT}{dataframe. includes all prospect input parameters}

\item{CHL}{numeric. Chlorophyll content (microg.cm-2)}

\item{CAR}{numeric. Carotenoid content (microg.cm-2)}

\item{ANT}{numeric. Anthocyain content (microg.cm-2)}

\item{BROWN}{numeric. Brown pigment content (Arbitrary units)}

\item{EWT}{numeric. Equivalent Water Thickness (g.cm-2)}

\item{LMA}{numeric. Leaf Mass per Area (g.cm-2)}

\item{PROT}{numeric. protein content  (g.cm-2)}

\item{CBC}{numeric. NonProtCarbon-based constituent content (g.cm-2)}

\item{N}{numeric. Leaf structure parameter}

\item{alpha}{numeric. Solid angle for incident light at surface of leaf}

\item{fraction_brown}{numeric. fraction of brown vegetation (0-1)}

\item{BrownLOP}{dataframe. brown leaf optical properties, when available

(simulation of roughness)}
}
\value{
invisible
}
\description{
adjust prospect inputs and run prospect in preparation for 4SAIL or 4SAIL2
}
