% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_prosail_inversion.R
\name{apply_prosail_inversion}
\alias{apply_prosail_inversion}
\title{This function applies SVR model on raster data in order to estimate
vegetation biophysical properties}
\usage{
apply_prosail_inversion(
  raster_path,
  mask_path = NULL,
  hybrid_model,
  output_dir,
  band_names,
  selected_bands,
  options = NULL
)
}
\arguments{
\item{raster_path}{character. path for a raster file}

\item{mask_path}{character. path for binary mask defining ON (1) and OFF (0)
pixels in the raster}

\item{hybrid_model}{list. hybrid models produced from train_prosail_inversion
each element of the list corresponds to a set of hybrid models for a
vegetation parameter}

\item{output_dir}{character. path for directory where results are written}

\item{band_names}{character. spectral bands corresponding to the raster}

\item{selected_bands}{list. list of spectral bands to be selected from raster
(identified by name of vegetation parameter)}

\item{options}{\itemize{
\item multiplying_factor numeric. to be applied to reflectance in the raster
--> PROSAIL simulates reflectance between 0 and 1, and raster data expected
in the same range
\item tiling boolean. process per block (using raster) or per tile (using terra)
\item tile_size numeric. size of individual tiles in meters
\item maxRows numeric. nb of rows to be read from raster
\item progressBar boolean. if progress bar required
\item filetype character. driver for raster file
}}
}
\value{
bp_path list. list of path for output files corresponding to
biophysical properties
}
\description{
This function applies SVR model on raster data in order to estimate
vegetation biophysical properties
}
