% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SENSOR.R
\name{applySensorCharacteristics}
\alias{applySensorCharacteristics}
\title{reads spectral response from known sensor
spectral response from Sentinel-2 is already defined}
\usage{
applySensorCharacteristics(wvl, InRefl, SRF)
}
\arguments{
\item{wvl}{numeric. spectral sampling of the input spectral data}

\item{InRefl}{numeric. input spectral data (unique sample)}

\item{SRF}{list. Information about spectral response Spectral Bands of the sensor and Original Bands for which SRF is defined}
}
\value{
OutRefl numeric. Output spectral data, sensor resolution
}
\description{
reads spectral response from known sensor
spectral response from Sentinel-2 is already defined
}
