% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SpectralIndices.R
\name{AUC}
\alias{AUC}
\title{This function computes Area under curve for continuum removed reflectances
See Malenovský et al. (2013) for details
http://dx.doi.org/10.1016/j.rse.2012.12.015}
\usage{
AUC(Refl, SensorBands, AUCminmax, ReflFactor = 1)
}
\arguments{
\item{Refl}{RasterBrick, RasterStack or list. Raster bands in the order of SensorBands.}

\item{SensorBands}{numeric. vector containing central wavelength for each spectral band in the image}

\item{AUCminmax}{list. wavelengths of lower and upper boundaries ('CRmin' and 'CRmax')}

\item{ReflFactor}{numeric. multiplying factor used to write reflectance in image (==10000 for S2)}
}
\value{
AUCval raster
}
\description{
This function computes Area under curve for continuum removed reflectances
See Malenovský et al. (2013) for details
http://dx.doi.org/10.1016/j.rse.2012.12.015
}
