% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT.R
\name{calctav}
\alias{calctav}
\title{computation of transmissivity of a dielectric plane surface,
averaged over all directions of incidence and over all polarizations.}
\usage{
calctav(alpha, nr)
}
\arguments{
\item{alpha}{numeric. Maximum incidence angle defining the solid angle of incident light}

\item{nr}{numeric. refractive index}
}
\value{
numeric. Transmissivity of a dielectric plane surface
}
\description{
computation of transmissivity of a dielectric plane surface,
averaged over all directions of incidence and over all polarizations.
}
