% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL.R
\name{campbell}
\alias{campbell}
\title{Computes the leaf angle distribution function value (freq)
Ellipsoidal distribution function caracterised by the average leaf
inclination angle in degree (ala)
Campbell 1986}
\usage{
campbell(ala)
}
\arguments{
\item{ala}{average leaf angle}
}
\value{
foliar_distrib list. lidf and litab
}
\description{
Computes the leaf angle distribution function value (freq)
Ellipsoidal distribution function caracterised by the average leaf
inclination angle in degree (ala)
Campbell 1986
}
