% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_brown_lop.R
\name{check_brown_lop}
\alias{check_brown_lop}
\title{check if brown leaf optical properties are correctly defined}
\usage{
check_brown_lop(brown_lop, lambda, input_prospect)
}
\arguments{
\item{brown_lop}{dataframe. should include wvl, Reflectance & Transmittance}

\item{lambda}{numeric. spectral bands corresponding to data to simulate}

\item{input_prospect}{dataframe. includes all prospect input parameters}
}
\value{
invisible
}
\description{
check if brown leaf optical properties are correctly defined
}
