% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL.R
\name{check_BrownLOP}
\alias{check_BrownLOP}
\title{check if brown leaf optical properties are correctly defined}
\usage{
check_BrownLOP(BrownLOP, lambda, Input_PROSPECT)
}
\arguments{
\item{BrownLOP}{dataframe. should include wvl, Reflectance & Transmittance}

\item{lambda}{numeric. spectral bands corresponding to data to simulate}

\item{Input_PROSPECT}{dataframe. includes all prospect input parameters}
}
\value{
invisible
}
\description{
check if brown leaf optical properties are correctly defined
}
