% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL.R
\name{check_SpectralSampling}
\alias{check_SpectralSampling}
\title{check if spectral sampling is identical between SpecPROSPECT, SpecSOIL, SpecATM}
\usage{
check_SpectralSampling(SpecPROSPECT, SpecSOIL, SpecATM)
}
\arguments{
\item{SpecPROSPECT}{list. Includes optical constants required for PROSPECT}

\item{SpecSOIL}{list. Includes either dry soil and wet soil, or a unique soil sample if the psoil parameter is not inverted}

\item{SpecATM}{list. direct and diffuse radiation for clear conditions}
}
\value{
invisible
}
\description{
check if spectral sampling is identical between SpecPROSPECT, SpecSOIL, SpecATM
}
