% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT.R
\name{Complete_Input_PROSPECT}
\alias{Complete_Input_PROSPECT}
\title{Complete the list of PROSPECT parameters with default values}
\usage{
Complete_Input_PROSPECT(Input_PROSPECT, Parm2Add, ExpectedParms)
}
\arguments{
\item{Input_PROSPECT}{input parameters sent to PROSPECT by user}

\item{Parm2Add}{Parameters to be added to input parameters}

\item{ExpectedParms}{full set of parameters expected to run PROSPECT}
}
\value{
Input_PROSPECT
}
\description{
Complete the list of PROSPECT parameters with default values
}
