% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_albedo.R
\name{compute_albedo}
\alias{compute_albedo}
\title{Computes albedo
borrowed from python package PROSAIL developed by Jose Gomez Dans}
\usage{
compute_albedo(rsdstar, rddstar, tts, SpecATM_Sensor, PAR_range = c(400, 2400))
}
\arguments{
\item{rsdstar}{numeric. reflectance from direct light}

\item{rddstar}{numeric. reflectance from diffuse light}

\item{tts}{numeric. Solar zenith angle}

\item{SpecATM_Sensor}{list. direct and diffuse radiation for clear conditions}

\item{PAR_range}{numeric. range (in nm) of spectral domain to integrate for
computation of albedo}
}
\value{
albedo numeric. albedo
}
\description{
The direct and diffuse light are taken into account as proposed by:
Francois et al. (2002) Conversion of 400-1100 nm vegetation albedo
measurements into total shortwave broadband albedo using a canopy
radiative transfer model, Agronomie
Es = direct
Ed = diffuse
}
