% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fAPAR.R
\name{compute_fAPAR}
\alias{compute_fAPAR}
\title{Computes fraction of absorbed photosyntehtically active radiation fAPAR}
\usage{
compute_fAPAR(abs_dir, abs_hem, tts, SpecATM_Sensor, PAR_range = c(400, 700))
}
\arguments{
\item{abs_dir}{numeric. fraction of direct light absorbed}

\item{abs_hem}{numeric. fraction of diffuse light absorbed}

\item{tts}{numeric. Solar zenith angle}

\item{SpecATM_Sensor}{list. direct and diffuse radiation for clear conditions}

\item{PAR_range}{numeric. range (in nm) of spectral domain to integrate for
computation of fAPAR}
}
\value{
fAPAR numeric. fAPAR
}
\description{
The direct and diffuse light are taken into account as proposed by:
Francois et al. (2002) Conversion of 400-1100 nm vegetation albedo
measurements into total shortwave broadband albedo using a canopy
radiative transfer model, Agronomie
Es = direct
Ed = diffuse
}
