% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SpectralIndices.R
\name{ComputeCR}
\alias{ComputeCR}
\title{This function computes continuum removal value for a spectral band of interest,
based on lower and upper wavelengths corresponding to boundaries of the continuum}
\usage{
ComputeCR(WLmin, WLmax, WLtarget, boundaries, target)
}
\arguments{
\item{WLmin}{numeric. wavelength of the spectral band corresponding to minimum boundary}

\item{WLmax}{numeric. wavelength of the spectral band corresponding to maximum boundary}

\item{WLtarget}{numeric. wavelength of the spectral band for which CR is computed}

\item{boundaries}{list. raster objects corresponding to minimum and maximum wavelengths}

\item{target}{list. raster object corresponding target wavelength}
}
\value{
CR list. raster object corresponding to continuum removed value
}
\description{
This function computes continuum removal value for a spectral band of interest,
based on lower and upper wavelengths corresponding to boundaries of the continuum
}
