% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SpectralIndices.R
\name{ComputeSpectralIndices_fromExpression}
\alias{ComputeSpectralIndices_fromExpression}
\title{this function produces a spectral index from an expression defining a spectral index}
\usage{
ComputeSpectralIndices_fromExpression(
  Refl,
  SensorBands,
  ExpressIndex,
  listBands,
  ReflFactor = 1,
  NameIndex = NULL
)
}
\arguments{
\item{Refl}{RasterBrick, RasterStack or list. Raster bands in the order of SensorBands.}

\item{SensorBands}{numeric. wavelength in nanometers of the spectral bands of Refl.}

\item{ExpressIndex}{character. expression corresponding to the spectral index to compute}

\item{listBands}{list. list of spectral bands defined in the 'ExpressIndex' variable}

\item{ReflFactor}{numeric. multiplying factor used to write reflectance in image (==10000 for S2)}

\item{NameIndex}{character. name for the index to be computed, provided in the raster layer}
}
\value{
numeric. band numbers of original sensor corresponding to S2
}
\description{
this function produces a spectral index from an expression defining a spectral index
}
