% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SpectralIndices.R
\name{ComputeSpectralIndices_HS}
\alias{ComputeSpectralIndices_HS}
\title{this function aims at computing spectral indices from Sensor reflectance data.
it computes the spectral indices based on their computation with Sentinel-2
and assumes that the bands of the S2 data follow this order
wavelength	= {496.6, 560.0, 664.5, 703.9, 740.2, 782.5, 835.1, 864.8, 1613.7, 2202.4}
Full description of the indices can be found here:
https://www.sentinel-hub.com/eotaxonomy/indices}
\usage{
ComputeSpectralIndices_HS(Refl, SensorBands, Sel_Indices = "ALL")
}
\arguments{
\item{Refl}{numeric. Reflectance dataset defined in matrix}

\item{SensorBands}{numeric. wavelength of the spectral bands corresponding to the spectral index}

\item{Sel_Indices}{list. list of spectral indices to be computed}
}
\value{
list. includes
\itemize{
\item SpectralIndices: List of spectral indices computed from the reflectance initially provided
\item listIndices: list of spectral indices computable with the function
}
}
\description{
this function aims at computing spectral indices from Sensor reflectance data.
it computes the spectral indices based on their computation with Sentinel-2
and assumes that the bands of the S2 data follow this order
wavelength	= {496.6, 560.0, 664.5, 703.9, 740.2, 782.5, 835.1, 864.8, 1613.7, 2202.4}
Full description of the indices can be found here:
https://www.sentinel-hub.com/eotaxonomy/indices
}
