% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL.R
\name{ConservativeScattering}
\alias{ConservativeScattering}
\title{computes conservative scattering conditions}
\usage{
ConservativeScattering(m, lai, att, sigb, ks, ko, sf, sb, vf, vb, tss, too)
}
\arguments{
\item{m}{numeric.}

\item{lai}{numeric. Leaf Area Index}

\item{att}{numeric.}

\item{sigb}{numeric.}

\item{ks}{numeric.}

\item{ko}{numeric.}

\item{sf}{numeric.}

\item{sb}{numeric.}

\item{vf}{numeric.}

\item{vb}{numeric.}

\item{tss}{numeric.}

\item{too}{numeric.}
}
\value{
list. tdd, rdd, tsd, rsd, tdo, rdo, rsod
}
\description{
computes conservative scattering conditions
}
