% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_function_rmse_prosail.R
\name{cost_function_rmse_prosail}
\alias{cost_function_rmse_prosail}
\title{Value of the cost criterion to minimize during PROSAIL inversion}
\usage{
cost_function_rmse_prosail(refl_mes, refl_mod, xprior, prior_info = NULL)
}
\arguments{
\item{refl_mes}{numeric. Measured surface reflectance}

\item{refl_mod}{numeric. Simulated surface reflectance}

\item{xprior}{list. values of the parameters for which prior information is
provided}

\item{prior_info}{list. prior mean, sd and weight of parameters defined as
xprior modulate its importance}
}
\value{
cost
}
\description{
Value of the cost criterion to minimize during PROSAIL inversion
}
