% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT_Inversion.R
\name{CostVal_RMSE}
\alias{CostVal_RMSE}
\title{Value of the cost criterion to minimize during PROSPECT inversion}
\usage{
CostVal_RMSE(RT, Refl, Tran)
}
\arguments{
\item{RT}{list. Simulated reflectance and transmittance}

\item{Refl}{numeric. Reflectance on which PROSPECT ins inverted}

\item{Tran}{numeric. Transmittance on which PROSPECT ins inverted}
}
\value{
fc sum of squared difference between simulated and measured leaf optical properties
}
\description{
Value of the cost criterion to minimize during PROSPECT inversion
}
