% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL_Inversion.R
\name{CostVal_RMSE_PROSAIL_PRIOR}
\alias{CostVal_RMSE_PROSAIL_PRIOR}
\title{Value of the cost criterion to minimize during PROSAIL inversion
This function includes prior information (mean & SD) about a selection of variables}
\usage{
CostVal_RMSE_PROSAIL_PRIOR(brfMES, brfMOD)
}
\arguments{
\item{brfMES}{numeric. Measured bidirectional reflectance}

\item{brfMOD}{numeric. Simulated bidirectional reflectance}
}
\value{
fc sum of squared difference between simulated and measured leaf optical properties
}
\description{
Value of the cost criterion to minimize during PROSAIL inversion
This function includes prior information (mean & SD) about a selection of variables
}
