% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SpectralIndices.R
\name{CR_WL}
\alias{CR_WL}
\title{This function extracts boundaries to be used to compute continuum from reflectance data}
\usage{
CR_WL(Refl, SensorBands, CRbands, ReflFactor = 1)
}
\arguments{
\item{Refl}{RasterBrick, RasterStack or list. Raster bands in the order of SensorBands.}

\item{SensorBands}{numeric. vector containing central wavelength for each spectral band in the image}

\item{CRbands}{list. list of spectral bands (central wavelength) including CRmin and CRmax}

\item{ReflFactor}{numeric. multiplying factor used to write reflectance in image (==10000 for S2)}
}
\value{
outlier_IQR numeric. band numbers of original sensor corresponding to S2
}
\description{
This function extracts boundaries to be used to compute continuum from reflectance data
}
