% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcum.R
\name{dcum}
\alias{dcum}
\title{dcum function}
\usage{
dcum(a, b, t)
}
\arguments{
\item{a}{numeric. controls the average leaf slope}

\item{b}{numeric. controls the distribution's bimodality}

\item{t}{numeric. angle}
}
\value{
f
}
\description{
dcum function
}
