% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{adjust_PROSPECT_2_SAIL}
\alias{apply_noise_AddMult}
\alias{Apply_Noise_LUT}
\alias{Apply_prosail_inversion}
\alias{applySensorCharacteristics}
\alias{check_BrownLOP}
\alias{check_SpectralSampling}
\alias{Compute_albedo}
\alias{Compute_BRF}
\alias{Compute_fAPAR}
\alias{Compute_SRF}
\alias{ConservativeScattering}
\alias{CostVal_RMSE_PROSAIL}
\alias{Generate_LUT_4SAIL}
\alias{get_HDR_name}
\alias{get_default_LUT_input}
\alias{get_atbd_LUT_input}
\alias{get_InputPROSAIL}
\alias{get_S2geometry}
\alias{get_S2geometry_from_SAFE}
\alias{get_S2geometry_from_THEIA}
\alias{GetRadiometry}
\alias{Invert_PROSAIL}
\alias{Jfunc1}
\alias{Jfunc2}
\alias{Jfunc3}
\alias{Jfunc4}
\alias{Merit_RMSE_PROSAIL}
\alias{NonConservativeScattering}
\alias{PrepareSensorSimulation}
\alias{PRO4SAIL}
\alias{PROSAIL_Hybrid_Apply}
\alias{PROSAIL_Hybrid_Train}
\alias{read_ENVI_header}
\alias{WhichParmPrior}
\alias{WhichParameters2Invert}
\alias{write_ENVI_header}
\title{Deprecated functions in prosail}
\usage{
adjust_PROSPECT_2_SAIL(
  SAILversion,
  Spec_Sensor,
  Input_PROSPECT,
  CHL,
  CAR,
  ANT,
  BROWN,
  EWT,
  LMA,
  PROT,
  CBC,
  N,
  alpha,
  fraction_brown,
  BrownLOP = NULL
)

apply_noise_AddMult(BRF_LUT, AdditiveNoise = 0.01, MultiplicativeNoise = 0.02)

Apply_Noise_LUT(LUT, NoiseLevel, NoiseType = "relative")

Apply_prosail_inversion(
  raster_path,
  HybridModel,
  PathOut,
  SelectedBands,
  bandname,
  MaskRaster = NULL,
  MultiplyingFactor = 10000,
  maxRows = 100,
  bigRaster = FALSE,
  progressBar = TRUE,
  filetype = "GTiff"
)

applySensorCharacteristics(wvl, InRefl, SRF)

check_BrownLOP(BrownLOP, lambda, Input_PROSPECT)

check_SpectralSampling(SpecPROSPECT, SpecSOIL, SpecATM)

Compute_albedo(rsdstar, rddstar, tts, SpecATM_Sensor, PAR_range = c(400, 2400))

Compute_BRF(rdot, rsot, tts, SpecATM_Sensor, skyl = NULL)

Compute_fAPAR(abs_dir, abs_hem, tts, SpecATM_Sensor, PAR_range = c(400, 700))

Compute_SRF(wvl, FWHM, SensorName = "user_defined")

ConservativeScattering(m, lai, att, sigb, ks, ko, sf, sb, vf, vb, tss, too)

CostVal_RMSE_PROSAIL(
  brfMES,
  brfMOD,
  xprior,
  PriorInfoMean = NULL,
  PriorInfoSD = NULL,
  WeightPrior = 0.01
)

Generate_LUT_4SAIL(
  InputPROSAIL,
  SpecPROSPECT,
  SpecSOIL,
  SpecATM,
  BandNames = NULL,
  SAILversion = "4SAIL",
  BrownLOP = NULL
)

get_HDR_name(ImPath)

get_default_LUT_input(
  TypeDistrib = NULL,
  GaussianDistrib = NULL,
  minval = NULL,
  maxval = NULL
)

get_atbd_LUT_input(nbSamples = 2000, GeomAcq = NULL, Codist_LAI = TRUE)

get_InputPROSAIL(
  atbd = FALSE,
  GeomAcq = NULL,
  Codist_LAI = TRUE,
  minval = NULL,
  maxval = NULL,
  TypeDistrib = NULL,
  GaussianDistrib = NULL,
  ParmSet = NULL,
  nbSamples = 2000,
  verbose = FALSE
)

get_S2geometry(MTD_TL_xml, verbose = FALSE)

get_S2geometry_from_SAFE(s2xml)

get_S2geometry_from_THEIA(s2xml)

GetRadiometry(
  SensorName = "user_defined",
  SpectralProps = NULL,
  Path_SensorResponse = "./",
  SaveSRF = TRUE
)

Invert_PROSAIL(
  brfMES,
  InitialGuess = NULL,
  LowerBound,
  UpperBound,
  SpecPROSPECT_Sensor,
  SpecATM_Sensor,
  SpecSOIL_Sensor,
  TypeLidf,
  ParmSet,
  MeritFunction = "Merit_RMSE_PROSAIL",
  PriorInfoMean = NULL,
  PriorInfoSD = NULL,
  WeightPrior = 0.01
)

Jfunc1(k, l, t)

Jfunc2(k, l, t)

Jfunc3(k, l, t)

Jfunc4(m, t)

Merit_RMSE_PROSAIL(
  xinit,
  parms_xinit,
  brfMES,
  SpecPROSPECT_Sensor,
  SpecSOIL_Sensor,
  SpecATM_Sensor,
  Parms2Estimate,
  Parm2Set = NULL,
  ParmSet = NULL,
  InVar,
  TypeLidf,
  PriorInfoMean = NULL,
  PriorInfoSD = NULL,
  Parms2Prior = NULL,
  WeightPrior = 0.01
)

NonConservativeScattering(m, lai, att, sigb, ks, ko, sf, sb, vf, vb, tss, too)

PrepareSensorSimulation(SpecPROSPECT, SpecSOIL, SpecATM, SRF)

PRO4SAIL(
  Spec_Sensor = NULL,
  Input_PROSPECT = NULL,
  N = 1.5,
  CHL = 40,
  CAR = 8,
  ANT = 0,
  BROWN = 0,
  EWT = 0.01,
  LMA = NULL,
  PROT = 0,
  CBC = 0,
  alpha = 40,
  TypeLidf = 2,
  LIDFa = 60,
  LIDFb = NULL,
  lai = 3,
  q = 0.1,
  tts = 30,
  tto = 0,
  psi = 60,
  rsoil = NULL,
  fraction_brown = 0,
  diss = 0,
  Cv = 1,
  Zeta = 1,
  SAILversion = "4SAIL",
  BrownLOP = NULL
)

PROSAIL_Hybrid_Apply(RegressionModels, Refl, progressBar = FALSE)

PROSAIL_Hybrid_Train(
  BRF_LUT,
  InputVar,
  nbEnsemble = 20,
  WithReplacement = FALSE,
  method = "liquidSVM",
  verbose = FALSE,
  progressBar = FALSE
)

read_ENVI_header(HDRpath)

WhichParmPrior(PriorInfoMean, PriorInfoSD)

WhichParameters2Invert(InitialGuess, LowerBound, UpperBound, ParmSet)

write_ENVI_header(HDR, HDRpath)
}
\arguments{
\item{SAILversion, Spec_Sensor, Input_PROSPECT, lambda}{See the new functions that replace the old ones}

\item{CHL, CAR, ANT, BROWN, EWT, LMA, PROT, CBC, N, alpha, fraction_brown, BrownLOP}{See the new functions that replace the old ones}

\item{BRF_LUT, AdditiveNoise, MultiplicativeNoise}{See the new functions that replace the old ones}

\item{raster_path, HybridModel, PathOut, SelectedBands, bandname, MaskRaster}{See the new functions that replace the old ones}

\item{MultiplyingFactor, maxRows, bigRaster, progressBar, filetype}{See the new functions that replace the old ones}

\item{SpecATM_Sensor, SpecSOIL_Sensor, TypeLidf, ParmSet, MeritFunction}{See the new functions that replace the old ones}

\item{rdot, rsot, tts, skyl, rsdstar, rddstar, PAR_range, abs_dir, abs_hem}{See the new functions that replace the old ones}

\item{m, lai, att, sigb, ks, ko, sf, sb, vf, vb, tss, too, t, k, l}{See the new functions that replace the old ones}

\item{brfMES, brfMOD, xprior, PriorInfoMean, PriorInfoSD, WeightPrior}{See the new functions that replace the old ones}

\item{InputPROSAIL, SpecPROSPECT, SpecSOIL, SpecATM, BandNames}{See the new functions that replace the old ones}

\item{TypeDistrib, GaussianDistrib, minval, maxval}{See the new functions that replace the old ones}

\item{nbSamples, GeomAcq, Codist_LAI}{See the new functions that replace the old ones}

\item{s2xml, MTD_TL_xml, verbose, atbd, ImPath, InRefl, SRF}{See the new functions that replace the old ones}

\item{SpectralProps, Path_SensorResponse, SaveSRF}{See the new functions that replace the old ones}

\item{InitialGuess, LowerBound, UpperBound, SpecPROSPECT_Sensor}{See the new functions that replace the old ones}

\item{xinit, parms_xinit, Parms2Estimate, Parm2Set, InVar, Parms2Prior}{See the new functions that replace the old ones}

\item{LIDFa, LIDFb, q, tto, psi, rsoil, diss, Cv, Zeta}{See the new functions that replace the old ones}

\item{InputVar, nbEnsemble, WithReplacement, method, RegressionModels, Refl}{See the new functions that replace the old ones}

\item{HDR, HDRpath, wvl, FWHM, SensorName, LUT, NoiseLevel, NoiseType}{See the new functions that replace the old ones}
}
\description{
These functions are provided for compatibility with older versions of prosail but are deprecated.
}
