% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dladgen.R
\name{dladgen}
\alias{dladgen}
\title{Computes the leaf angle distribution function value (freq)}
\usage{
dladgen(a, b)
}
\arguments{
\item{a}{controls the average leaf slope}

\item{b}{controls the distribution's bimodality
LIDF type           a         b
Planophile       1          0
Erectophile    -1         0
Plagiophile       0          -1
Extremophile     0          1
Spherical         -0.35 -0.15
Uniform        0     0
requirement: |lidf_a| + |lidf_b| < 1}
}
\value{
foliar_distrib list. lidf and litab
}
\description{
\subsection{Using the original bimodal distribution function initially proposed in SAIL
References}{

(Verhoef1998) Verhoef, Wout. Theory of radiative transfer models applied
in optical remote sensing of vegetation canopies.
Nationaal Lucht en Ruimtevaartlaboratorium, 1998.
http://library.wur.nl/WebQuery/clc/945481.
}
}
