% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT_Inversion.R
\name{FitSpectralData}
\alias{FitSpectralData}
\title{This function adapts SpecPROSPECT accordingly to experimental data
or to a spectral domain defined by UserDomain}
\usage{
FitSpectralData(SpecPROSPECT, lambda, Refl = NULL, Tran = NULL,
  UserDomain = NULL)
}
\arguments{
\item{SpecPROSPECT}{list. Includes optical constants
refractive index, specific absorption coefficients and corresponding spectral bands}

\item{lambda}{numeric. Spectral bands corresponding to experimental data}

\item{Refl}{numeric. Measured reflectance data}

\item{Tran}{numeric. Measured Transmittance data}

\item{UserDomain}{numeric. Lower and upper bounds for domain of interest (optional)}
}
\value{
res list including spectral properties at the new resolution
}
\description{
This function adapts SpecPROSPECT accordingly to experimental data
or to a spectral domain defined by UserDomain
}
