% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourSAIL.R
\name{fourSAIL}
\alias{fourSAIL}
\title{Performs PROSAIL simulation based on a set of combinations of
input parameters}
\usage{
fourSAIL(
  lop,
  TypeLidf = 2,
  LIDFa = 60,
  LIDFb = NULL,
  lai = 3,
  q = 0.1,
  tts = 30,
  tto = 0,
  psi = 60,
  rsoil = NULL
)
}
\arguments{
\item{lop}{list. leaf optical properties
(reflectance and transmittance) and corresponding spectral bands}

\item{TypeLidf}{numeric. Type of leaf inclination distribution function}

\item{LIDFa}{numeric.
if TypeLidf ==1, controls the average leaf slope
if TypeLidf ==2, controls the average leaf angle}

\item{LIDFb}{numeric.
if TypeLidf ==1, controls the distribution's bimodality
if TypeLidf ==2, unused}

\item{lai}{numeric. Leaf Area Index}

\item{q}{numeric. Hot Spot parameter}

\item{tts}{numeric. Sun zeith angle}

\item{tto}{numeric. Observer zeith angle}

\item{psi}{numeric. Azimuth Sun / Observer}

\item{rsoil}{numeric. Soil reflectance}
}
\value{
list. rdot,rsot,rddt,rsdt
rdot: hemispherical-directional reflectance factor in viewing direction
rsot: bi-directional reflectance factor
rsdt: directional-hemispherical reflectance factor for solar incident flux
rddt: bi-hemispherical reflectance factor
fcover: Fraction of green Vegetation Cover (= 1 - beam transmittance in
the target-view path)
abs_dir: canopy absorptance for direct solar incident flux
abs_hem: canopy absorptance for hemispherical diffuse incident flux
rsdstar: contribution of direct solar incident flux to albedo
rddstar: contribution of hemispherical diffuse incident flux to albedo
}
\description{
Performs PROSAIL simulation based on a set of combinations of
input parameters
}
