% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL.R
\name{fourSAIL2}
\alias{fourSAIL2}
\title{Performs PRO4SAIL2 simulation based on a set of combinations of input parameters}
\usage{
fourSAIL2(
  leafgreen,
  leafbrown,
  TypeLidf = 2,
  LIDFa = NULL,
  LIDFb = NULL,
  lai = NULL,
  hot = NULL,
  tts = NULL,
  tto = NULL,
  psi = NULL,
  rsoil = NULL,
  fraction_brown = 0.5,
  diss = 0.5,
  Cv = 1,
  Zeta = 1
)
}
\arguments{
\item{leafgreen}{list. includes relfectance and transmittance for vegetation #1 (e.g. green vegetation)}

\item{leafbrown}{list. includes relfectance and transmittance for vegetation #2 (e.g. brown vegetation)}

\item{TypeLidf}{numeric. Type of leaf inclination distribution function}

\item{LIDFa}{numeric.
if TypeLidf ==1, controls the average leaf slope
if TypeLidf ==2, controls the average leaf angle}

\item{LIDFb}{numeric.
if TypeLidf ==1, controls the distribution's bimodality
if TypeLidf ==2, unused}

\item{lai}{numeric. Leaf Area Index}

\item{hot}{numeric. Hot Spot parameter = ratio of the correlation length of leaf projections in the horizontal plane and the canopy height (doi:10.1016/j.rse.2006.12.013)}

\item{tts}{numeric. Sun zeith angle}

\item{tto}{numeric. Observer zeith angle}

\item{psi}{numeric. Azimuth Sun / Observer}

\item{rsoil}{numeric. Soil reflectance}

\item{fraction_brown}{numeric. Fraction of brown leaf area}

\item{diss}{numeric. Layer dissociation factor}

\item{Cv}{numeric. vertical crown cover percentage
= \% ground area covered with crowns as seen from nadir direction}

\item{Zeta}{numeric. Tree shape factor
= ratio of crown diameter to crown height}
}
\value{
list. rdot,rsot,rddt,rsdt
rdot: hemispherical-directional reflectance factor in viewing direction
rsot: bi-directional reflectance factor
rsdt: directional-hemispherical reflectance factor for solar incident flux
rddt: bi-hemispherical reflectance factor
alfast: canopy absorptance for direct solar incident flux
alfadt: canopy absorptance for hemispherical diffuse incident flux
}
\description{
Performs PRO4SAIL2 simulation based on a set of combinations of input parameters
}
