% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_lut_surf_refl.R
\name{generate_lut_surf_refl}
\alias{generate_lut_surf_refl}
\title{generates a LUT of surface reflectance based on a table of input variables for prosail model}
\usage{
generate_lut_surf_refl(
  input_prosail,
  spec_prospect,
  spec_soil,
  spec_atm,
  band_names = NULL,
  SAILversion = "4SAIL",
  brown_lop = NULL
)
}
\arguments{
\item{input_prosail}{list. PROSAIL input variables}

\item{spec_prospect}{list. Includes optical constants required for PROSPECT}

\item{spec_soil}{list. Includes either a set of OSSL library reflectance
spectra, reflectance spectra from S2 ATBD v2, or minimum reflectance and maximum reflectance}

\item{spec_atm}{list. Includes direct and diffuse radiation for
clear conditions}

\item{band_names}{character. Name of the spectral bands of the sensor}

\item{SAILversion}{character. choose between 4SAIL and 4SAIL2}

\item{brown_lop}{list. Defines optical properties for brown vegetation
\itemize{
\item WVL, Reflectance, Transmittance
\item Set to NULL if use PROSPECT to generate it
}}
}
\value{
surf_refl numeric. matrix of surf_refl corresponding to input_prosail
}
\description{
generates a LUT of surface reflectance based on a table of input variables for prosail model
}
\examples{
input_prosail <- get_input_prosail(atbd = TRUE, nb_samples = 100)
refl_lut <- generate_lut_surf_refl(input_prosail = input_prosail,
                                   spec_prospect = prosail::spec_prospect_full_range,
                                   spec_soil = prosail::spec_soil_atbd_v2,
                                   spec_atm = prosail::spec_atm)

}
