% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL_LUT.R
\name{get_atbd_LUT_input}
\alias{get_atbd_LUT_input}
\title{This function provides a LUT corresponding to PROSAIL input parameters to
reproduce the distribution of parameters used in the ATBD}
\usage{
get_atbd_LUT_input(nbSamples = 2000, GeomAcq = NULL, Codist_LAI = TRUE)
}
\arguments{
\item{nbSamples}{numeric. number of samples to generate}

\item{GeomAcq}{list. min and max values for tts, tto and psi}

\item{Codist_LAI}{boolean. set TYRUE if codistribution with LAI accounted for}
}
\value{
InputPROSAIL list. list of PROSAIL input parameters
}
\description{
This function provides a LUT corresponding to PROSAIL input parameters to
reproduce the distribution of parameters used in the ATBD
}
