% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_atbd_v3_lut_input.R
\name{get_atbd_v3_lut_input}
\alias{get_atbd_v3_lut_input}
\title{This function provides a LUT corresponding to PROSAIL input parameters to
reproduce the distribution of parameters used in the ATBD}
\usage{
get_atbd_v3_lut_input(
  nb_samples = 2000,
  geom_acq = NULL,
  codistribution_lai = TRUE
)
}
\arguments{
\item{nb_samples}{numeric. number of samples to generate}

\item{geom_acq}{list. min and max values for tts, tto and psi}

\item{codistribution_lai}{boolean. set TYRUE if codistribution with LAI accounted for}
}
\value{
input_prosail list. list of PROSAIL input parameters
}
\description{
This function provides a LUT corresponding to PROSAIL input parameters to
reproduce the distribution of parameters used in the ATBD
}
