% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SpectralIndices.R
\name{get_bands_close2s2}
\alias{get_bands_close2s2}
\title{this function identifies the bands of a given sensor with closest match to S2 bands}
\usage{
get_bands_close2s2(SensorBands, S2Bands)
}
\arguments{
\item{SensorBands}{numeric. wavelength in nanometer of the sensor of interest}

\item{S2Bands}{numeric or list. Named vector or list of spectral bands corresponding to Sentinel-2}
}
\value{
numeric. band numbers of original sensor corresponding to S2
}
\description{
this function identifies the bands of a given sensor with closest match to S2 bands
}
