% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bp_from_tile.R
\name{get_bp_from_tile}
\alias{get_bp_from_tile}
\title{computes biophysical properties from a raster subset based on ML regression model}
\usage{
get_bp_from_tile(
  individual_plot,
  plot_id,
  raster_path,
  bp_vars,
  hybrid_model,
  output_dir,
  selected_bands,
  band_names,
  site_name,
  mask_path = NULL,
  multiplying_factor = 10000,
  filetype = "GTiff",
  p = NULL
)
}
\arguments{
\item{individual_plot}{character. path for a raster file}

\item{plot_id}{character. path for a raster file}

\item{raster_path}{character. path for a raster file}

\item{bp_vars}{character. path for a raster file}

\item{hybrid_model}{list. hybrid models produced from train_prosail_inversion
each element of the list corresponds to a set of hybrid models for a
vegetation parameter}

\item{output_dir}{character. path for directory where results are written}

\item{selected_bands}{list. list of spectral bands to be selected from raster
(identified by name of vegetation parameter)}

\item{band_names}{character. spectral bands corresponding to the raster}

\item{site_name}{character. spectral bands corresponding to the raster}

\item{mask_path}{character. path for binary mask defining ON (1) and OFF (0)
pixels in the raster}

\item{multiplying_factor}{numeric. multiplying factor used to write
reflectance in the raster
--> PROSAIL simulates reflectance between 0 and 1, and raster data expected
in the same range}

\item{filetype}{character. driver for raster file}

\item{p}{object}
}
\value{
res character. path for output files corresponding to
biophysical properties
}
\description{
computes biophysical properties from a raster subset based on ML regression model
}
