% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SpectralIndices.R
\name{get_closest_bands}
\alias{get_closest_bands}
\title{this function identifies the bands of a given sensor with closest match to its spectral characteristics}
\usage{
get_closest_bands(SensorBands, listBands)
}
\arguments{
\item{SensorBands}{numeric. wavelength in nanometer of the sensor of interest}

\item{listBands}{numeric or list. Named vector or list of spectral bands corresponding to sensor}
}
\value{
numeric. band numbers of original sensor
}
\description{
this function identifies the bands of a given sensor with closest match to its spectral characteristics
}
