% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distribution_input_prosail.R
\name{get_distribution_input_prosail}
\alias{get_distribution_input_prosail}
\title{This function generates distribution of biophysical parameters used as input
parameters in prosail}
\usage{
get_distribution_input_prosail(
  minval = NULL,
  maxval = NULL,
  parm_set = NULL,
  nb_samples = 2000,
  type_distrib = NULL,
  mean = NULL,
  sd = NULL
)
}
\arguments{
\item{minval}{list. minimum value for a list of parameters randomly produced}

\item{maxval}{list. maximum value for a list of parameters randomly produced}

\item{parm_set}{list. parameters to be set to a given value}

\item{nb_samples}{numeric. Number of samples to be generated}

\item{type_distrib}{list. uniform/gaussian distribution to be applied.}

\item{mean}{list. mean value for parameters with gaussian distribution}

\item{sd}{list. standard deviation for parameters with gaussian distribution}
}
\value{
input_prosail list. list of nb_samples input parameters for
prosail simulation
}
\description{
This function generates distribution of biophysical parameters used as input
parameters in prosail
}
\examples{
distrib_inputs <- get_distribution_input_prosail(
    minval = data.frame('chl' = 5, 'car' = 1, 'lai' = 0.5),
    maxval = data.frame('chl' = 60, 'car' = 12, 'lai' = 6),
    parm_set = list('type_lidf' = 2, 'lidf_a' = 60),
    type_distrib = list('chl' = 'uniform',
                        'car' = 'uniform',
                        'lai' = 'gaussian'),
    mean = list('lai' = 3), sd = list('lai' = 2), nb_samples = 200)

}
