% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distribution_input_prosail2.R
\name{get_distribution_input_prosail2}
\alias{get_distribution_input_prosail2}
\title{This function generates distribution of biophysical parameters used as input
parameters in prosail}
\usage{
get_distribution_input_prosail2(
  minval,
  maxval,
  parm_set,
  nb_samples,
  type_distrib = NULL,
  mean = NULL,
  sd = NULL
)
}
\arguments{
\item{minval}{list. min val set for list of parameters randomly produced}

\item{maxval}{list. max val set for list of parameters randomly produced}

\item{parm_set}{list.Defines the parameters to be set to a given value}

\item{nb_samples}{numeric. Number of samples to be generated}

\item{type_distrib}{list. uniform / Gaussian distribution to be applied}

\item{mean}{list. mean value for parameters with Gaussian distribution}

\item{sd}{list. standard deviation for parameters with Gaussian distribution}
}
\value{
input_prosail list. input parameters for prosail simulation
}
\description{
This function generates distribution of biophysical parameters used as input
parameters in prosail
}
