% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid_bp.R
\name{get_grid_bp}
\alias{get_grid_bp}
\title{return vector grid corresponding to an area of interest}
\usage{
get_grid_bp(raster_path, cellsize = 10000)
}
\arguments{
\item{raster_path}{character. path for vector file defining aoi}

\item{cellsize}{numeric. square cell size in meters}
}
\value{
dsn_grid path for grid of aoi
}
\description{
return vector grid corresponding to an area of interest
}
