% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_HDR_name.R
\name{get_hdr_name}
\alias{get_hdr_name}
\title{get hdr name from image file name, assuming it is BIL format}
\usage{
get_hdr_name(image_path)
}
\arguments{
\item{image_path}{path of the image}
}
\value{
corresponding hdr
}
\description{
get hdr name from image file name, assuming it is BIL format
}
