% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_prosail.R
\name{get_input_prosail}
\alias{get_input_prosail}
\title{generate input_prosail, following
\itemize{
\item either distribution defined in ATBD
\item or distribution defined in user
}}
\usage{
get_input_prosail(
  atbd = FALSE,
  geom_acq = NULL,
  codistribution_lai = TRUE,
  minval = NULL,
  maxval = NULL,
  type_distrib = NULL,
  gaussian_distrib = NULL,
  parm_set = NULL,
  nb_samples = 2000,
  verbose = FALSE
)
}
\arguments{
\item{atbd}{boolean. TRUE to apply input parameter distribution from ATBD}

\item{geom_acq}{list. geom of acquisiton: min and max val for tts, tto & psi}

\item{codistribution_lai}{boolean. TRUE accounts for codistribution with LAI}

\item{minval}{list. min val for input parameters}

\item{maxval}{list. max val for input parameters}

\item{type_distrib}{list. Type of distribution: 'uniform' or 'gaussian'}

\item{gaussian_distrib}{list. Mean & SD for parms sampled with gaussian dist}

\item{parm_set}{list. list of input parameters set to a specific value}

\item{nb_samples}{numeric. number of samples in training LUT}

\item{verbose}{boolean. when set to TRUE, prints message if hyperparameter
adjustment performed during training}
}
\value{
input_prosail
}
\description{
generate input_prosail, following
\itemize{
\item either distribution defined in ATBD
\item or distribution defined in user
}
}
\examples{
input_prosail <- get_input_prosail(atbd = TRUE, nb_samples = 2000,
                                   geom_acq = data.frame('min' = c('tto' = 0,
                                                                   'tts' = 20,
                                                                   'psi' = 0),
                                                         'max' = c('tto' = 10,
                                                                   'tts' = 30,
                                                                   'psi' = 360)))
}
