% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL_LUT.R
\name{get_insitu_LUT_input}
\alias{get_insitu_LUT_input}
\title{Title}
\usage{
get_insitu_LUT_input(
  nbSamples = 2000,
  use_codist_LAI = TRUE,
  insitu_foliar = NULL
)
}
\arguments{
\item{insitu_foliar}{}
}
\description{
Title
}
