% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT_Inversion.R
\name{Get_Nprior}
\alias{Get_Nprior}
\title{This function defines a regression model to estimate N from R only or T only}
\usage{
Get_Nprior(SpecPROSPECT, lambda, Refl = NULL, Tran = NULL)
}
\arguments{
\item{SpecPROSPECT}{list. Includes optical constants
refractive index, specific absorption coefficients and corresponding spectral bands}

\item{lambda}{numeric. spectral bands corresponding to experimental data}

\item{Refl}{numeric. Measured reflectance data}

\item{Tran}{numeric. Measured Transmittance data}
}
\value{
Nprior vector corresponding to teh prior estimation of N based on R only or T only
}
\description{
This function defines a regression model to estimate N from R only or T only
}
