% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_list.R
\name{get_plot_list}
\alias{get_plot_list}
\title{get a list of plots defined in a vector file defined as input}
\usage{
get_plot_list(dsn, nbdigits = 3)
}
\arguments{
\item{dsn}{character. path for vector file}

\item{nbdigits}{numeric. number of digits to define plot names}
}
\value{
list of polygons
}
\description{
get a list of plots defined in a vector file defined as input
}
