% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_list_from_sf.R
\name{get_plot_list_from_sf}
\alias{get_plot_list_from_sf}
\title{get a list of plots defined in a vector file defined as input}
\usage{
get_plot_list_from_sf(sf_obj, nbdigits = 3)
}
\arguments{
\item{sf_obj}{list. sf object}

\item{nbdigits}{numeric. number of digits to define plot names}
}
\value{
list of polygons
}
\description{
get a list of plots defined in a vector file defined as input
}
