% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_s2_geometry_from_THEIA.R
\name{get_s2_geometry_from_THEIA}
\alias{get_s2_geometry_from_THEIA}
\title{returns geometry of acquisition for S2 image processed with MAJA
\itemize{
\item sza = list of sun zenith angle
\item saa = list of sun azimuth angle
\item vza = list of viewer zenith angle
\item vaa = list of viewer azimuth angle
}}
\usage{
get_s2_geometry_from_THEIA(s2xml)
}
\arguments{
\item{s2xml}{list. produced from reading XML metadata file with package XML}
}
\value{
List of S2 angles (sza, saa, vza, vaa)
}
\description{
returns geometry of acquisition for S2 image processed with MAJA
\itemize{
\item sza = list of sun zenith angle
\item saa = list of sun azimuth angle
\item vza = list of viewer zenith angle
\item vaa = list of viewer azimuth angle
}
}
