% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SENSOR.R
\name{get_S2geometry}
\alias{get_S2geometry}
\title{This function returns geometry of acquisition for S2 image
\itemize{
\item SZA = list of sun zenith angle
\item SAA = list of sun azimuth angle
\item VZA = list of viewer zenith angle
\item VAA = list of viewer azimuth angle
}}
\usage{
get_S2geometry(MTD_TL_xml)
}
\arguments{
\item{MTD_TL_xml}{character. Path for metadata file MTD_TL.xml}
}
\value{
List of S2 angles (SZA, SAA, VZA, VAA)
}
\description{
This function returns geometry of acquisition for S2 image
\itemize{
\item SZA = list of sun zenith angle
\item SAA = list of sun azimuth angle
\item VZA = list of viewer zenith angle
\item VAA = list of viewer azimuth angle
}
}
