% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_S2geometry_from_THEIA.R
\name{get_S2geometry_from_THEIA}
\alias{get_S2geometry_from_THEIA}
\title{returns geometry of acquisition for S2 image processed with MAJA
\itemize{
\item SZA = list of sun zenith angle
\item SAA = list of sun azimuth angle
\item VZA = list of viewer zenith angle
\item VAA = list of viewer azimuth angle
}}
\usage{
get_S2geometry_from_THEIA(s2xml)
}
\arguments{
\item{s2xml}{list. produced from reading XML metadata file with package XML}
}
\value{
List of S2 angles (SZA, SAA, VZA, VAA)
}
\description{
returns geometry of acquisition for S2 image processed with MAJA
\itemize{
\item SZA = list of sun zenith angle
\item SAA = list of sun azimuth angle
\item VZA = list of viewer zenith angle
\item VAA = list of viewer azimuth angle
}
}
