% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spec_sensor.R
\name{get_spec_sensor}
\alias{get_spec_sensor}
\title{Computes spectral sensor characteristics for optical constants required for
PROSPECT soil parameters as well as direct and diffuse radiation for clear
conditions}
\usage{
get_spec_sensor(
  sensor_name = "user_defined",
  spectral_properties = NULL,
  srf_path = NULL
)
}
\arguments{
\item{sensor_name}{character. name of the sensor. Either already defined, or
to be defined based on other inputs}

\item{spectral_properties}{list. Sensor spectral characteristics including
wl = central wavelength and fwhm = corresponding fwhm}

\item{srf_path}{character. Path where to get or save SRF}
}
\value{
srf list. Spectral response function, corresponding spectral bands,
and Original Bands
}
\description{
Computes spectral sensor characteristics for optical constants required for
PROSPECT soil parameters as well as direct and diffuse radiation for clear
conditions
}
