% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spectral_response_function.R
\name{get_spectral_response_function}
\alias{get_spectral_response_function}
\title{reads spectral response from known sensor
spectral response from Sentinel-2 is already defined}
\usage{
get_spectral_response_function(
  sensor_name = "user_defined",
  spectral_properties = NULL,
  srf_path = "./",
  save_srf = TRUE
)
}
\arguments{
\item{sensor_name}{character. name of the sensor}

\item{spectral_properties}{list. list of spectral properties including
wl = central wavelength and fwhm = corresponding fwhm}

\item{srf_path}{character. path for the file where srf should be
saved as CSV}

\item{save_srf}{boolean. Should srf be saved in file?}
}
\value{
srf list. Spectral response function, corresponding spectral bands,
and Original Bands
}
\description{
reads spectral response from known sensor
spectral response from Sentinel-2 is already defined
}
