% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_srf_sensor.R
\name{get_srf_sensor}
\alias{get_srf_sensor}
\title{reads spectral response from known sensor
spectral response from Sentinel-2 is already defined}
\usage{
get_srf_sensor(
  sensor_name = "user_defined",
  wl = NULL,
  fwhm = NULL,
  srf_path = "./",
  save_srf = TRUE
)
}
\arguments{
\item{sensor_name}{character. name of the sensor. predefined sensor SRF:
\itemize{
\item Sentinel_2, Sentinel_2A, Sentinel_2B, Sentinel_2C
\item Venus
\item Landsat_8, Landsat_9
\item MODIS
\item SPOT_6_7
\item PLEIADES_1A and 1B
\item user defined possible if providing central wl and FWHM for each band, or
srf file. see example in repository and reproduce proper format
https://github.com/jbferet/prosail/blob/master/data-raw/Sentinel_2A_Spectral_Response.csv
}}

\item{wl}{numeric. central wavelength of spectral bands}

\item{fwhm}{numeric. fwhm of spectral bands defined by wl}

\item{srf_path}{character. path for the file where srf should be saved as CSV}

\item{save_srf}{boolean. Should srf be saved in file?}
}
\value{
srf list. Spectral response function, corresponding spectral bands,
and Original Bands
}
\description{
reads spectral response from known sensor
spectral response from Sentinel-2 is already defined
}
\examples{
srf_2a <- get_srf_sensor(sensor_name = 'sentinel-2a')
srf_2b <- get_srf_sensor(sensor_name = 'Sentinel-2B')
srf_2c <- get_srf_sensor(sensor_name = 'S2C')
srf_venus <- get_srf_sensor(sensor_name = 'venus')
srf_modis <- get_srf_sensor(sensor_name = 'Modis')
srf_user_defined <- get_srf_sensor(sensor_name = 'mysensor',
                                   wl = seq(400, 1000, by = 100),
                                   fwhm = rep(x = 100, 7), save_srf = FALSE)

}
