% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_surf_refl.R
\name{get_surf_refl}
\alias{get_surf_refl}
\title{Computes surface reflectance based on outputs from PROSAIL & SZA}
\usage{
get_surf_refl(rdot, rsot, tts, spec_atm_sensor, skyl = NULL)
}
\arguments{
\item{rdot}{numeric. Hemispherical-directional refl factor viewing direction}

\item{rsot}{numeric. Bi-directional reflectance factor}

\item{tts}{numeric. Solar zenith angle}

\item{spec_atm_sensor}{list. direct & diffuse radiation for clear conditions}

\item{skyl}{numeric. values for skyl}
}
\value{
surf_refl numeric. surface reflectance
}
\description{
The direct and diffuse light are taken into account as proposed by:
Francois et al. (2002) Conversion of 400-1100 nm vegetation albedo
measurements into total shortwave broadband albedo using a canopy
radiative transfer model, Agronomie
Es = direct
Ed = diffuse
}
\examples{
refl <- prosail()
refl_surf <- get_surf_refl(refl$rdot, refl$rsot, tts = 30,
                           spec_atm_sensor = prosail::spec_atm)

}
