% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SENSOR.R
\name{GetRadiometry}
\alias{GetRadiometry}
\title{reads spectral response from known sensor
spectral response from Sentinel-2 is already defined}
\usage{
GetRadiometry(SensorName, Path_SensorResponse = NULL)
}
\arguments{
\item{SensorName}{character. name of the sensor}

\item{Path_SensorResponse}{character.}
}
\value{
SRF list. Spectral response function, corresponding spectral bands, and Original Bands
}
\description{
reads spectral response from known sensor
spectral response from Sentinel-2 is already defined
}
