% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_PROSPECT-data.R
\docType{data}
\name{Input_PROSPECT}
\alias{Input_PROSPECT}
\title{Outcomes of 10 combinations of input parameters for PROSPECT}
\format{A data frame with 10 rows and 6 variables:
\itemize{
\item \emph{CHL}: Chlorophyll content
\item \emph{CAR}: carotenoid content
\item \emph{ANT}: anthocyanin content
\item \emph{EWT}: Equivalent water thickness
\item \emph{LMA}: Leaf Mass per Area
\item \emph{N}: leaf structure parameter
}}
\usage{
Input_PROSPECT
}
\description{
corresponds to input parameters from PROSPECT-D:
}
\keyword{datasets}
