% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_prosail.R
\name{invert_prosail}
\alias{invert_prosail}
\title{Performs PROSAIL inversion using iterative optimization
in order to define estimate a set of user defined parameters}
\usage{
invert_prosail(
  brf_mes,
  initialization = NULL,
  lower_bound,
  upper_bound,
  spec_prospect_sensor,
  spec_atm_sensor,
  spec_soil_sensor,
  type_lidf,
  parm_set,
  merit_function = "merit_rmse_prosail",
  prior_info = NULL
)
}
\arguments{
\item{brf_mes}{numeric. measured BRF to be used during inversion}

\item{initialization}{list. Initial guess of parms_to_estimate}

\item{lower_bound}{list. Lower bound}

\item{upper_bound}{list. Upper bound}

\item{spec_prospect_sensor}{list. Includes optical constants
refractive index, specific absorption coefs and corresponding spectral bands}

\item{spec_atm_sensor}{list. direct & diffuse radiation for clear conditions}

\item{spec_soil_sensor}{list. includes either dry soil and wet soil,
or a unique soil sample if the psoil parameter is not inverted}

\item{type_lidf}{numeric. Type of leaf inclination distribution function}

\item{parm_set}{list. Parameters set to a fixed value by user}

\item{merit_function}{character. name of the merit function
with given criterion to minimize (default = RMSE)}

\item{prior_info}{list. prior mean, sd and weight of parameters defined as
xprior modulate its importance}
}
\value{
OutPROSPECT estimated values corresponding to parms_to_estimate
}
\description{
Performs PROSAIL inversion using iterative optimization
in order to define estimate a set of user defined parameters
}
