% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT_Inversion.R
\name{Invert_PROSPECT}
\alias{Invert_PROSPECT}
\title{Performs PROSPECT inversion using iterative optimization
in order to define estimate a set of user defined parameters}
\usage{
Invert_PROSPECT(
  SpecPROSPECT,
  Refl = NULL,
  Tran = NULL,
  Parms2Estimate = "ALL",
  ParmSet = NULL,
  PROSPECT_version = "D",
  MeritFunction = "Merit_RMSE_PROSPECT"
)
}
\arguments{
\item{SpecPROSPECT}{list. Includes optical constants
refractive index, specific absorption coefficients and corresponding spectral bands}

\item{Refl}{numeric. Measured reflectance data}

\item{Tran}{numeric. Measured Transmittance data}

\item{Parms2Estimate}{list. Parameters to estimate (can be 'ALL')}

\item{ParmSet}{list. Parameters set to a fixed value by user}

\item{PROSPECT_version}{character. '5', '5B', 'D', 'DB', 'PRO', 'PROB',}

\item{MeritFunction}{character. name of the function to be used as merit function
with given criterion to minimize (default = RMSE)}
}
\value{
OutPROSPECT estimated values corresponding to Parms2Estimate
}
\description{
Performs PROSPECT inversion using iterative optimization
in order to define estimate a set of user defined parameters
}
