% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SpectralIndices.R
\name{IQR_outliers}
\alias{IQR_outliers}
\title{This function computes interquartile range (IQR) criterion, which can be used
as a criterion for outlier detection}
\usage{
IQR_outliers(DistVal)
}
\arguments{
\item{DistVal}{numeric. vector of distribution of values}
}
\value{
outlier_IQR numeric. band numbers of original sensor corresponding to S2
}
\description{
This function computes interquartile range (IQR) criterion, which can be used
as a criterion for outlier detection
}
