% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL.R
\name{Jfunc1}
\alias{Jfunc1}
\title{J1 function with avoidance of singularity problem}
\usage{
Jfunc1(k, l, t)
}
\arguments{
\item{k}{numeric. Extinction coefficient for direct (solar or observer) flux}

\item{l}{numeric.}

\item{t}{numeric. Leaf Area Index}
}
\value{
Jout numeric.
}
\description{
J1 function with avoidance of singularity problem
}
