% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landsat_7-data.R
\docType{data}
\name{Landsat_7}
\alias{Landsat_7}
\title{Spectral response function (SRF) for Landsat-7}
\format{
A list including
\itemize{
\item \emph{Spectral_Response}: the SRF of each spectral band of the sensor for a
spectral domain compatible with PROSAIL
\item \emph{Spectral_Bands}: the spectral bands of the sensor (i.e. wavelength or
name of the band as character)
\item \emph{Original_Bands}: original spectral bands for which SRF is defined
(compatible with PROSAIL == 1 nm spectral sampling)
\item \emph{Central_WL}: Central wavelength (in nm) of the sensor spectral bands
}
}
\usage{
Landsat_7
}
\description{
Spectral response function (SRF) for Landsat-7
}
\keyword{datasets}
