% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merit_RMSE_PROSAIL.R
\name{merit_RMSE_PROSAIL}
\alias{merit_RMSE_PROSAIL}
\title{Merit function for PROSAIL inversion}
\usage{
merit_RMSE_PROSAIL(
  xinit,
  parms_xinit,
  brf_mes,
  SpecPROSPECT_Sensor,
  SpecSOIL_Sensor,
  SpecATM_Sensor,
  Parms2Estimate,
  input_prosail,
  TypeLidf,
  prior_info = NULL,
  Parms2Prior = NULL
)
}
\arguments{
\item{xinit}{numeric. Vector of input variables to estimate}

\item{parms_xinit}{character. name of parameters corresponding to xinit}

\item{brf_mes}{numeric. measured BRF}

\item{SpecPROSPECT_Sensor}{list. Includes optical constants for PROSPECT}

\item{SpecSOIL_Sensor}{list. Includes soil reflectance (either 2 references
for optimization of psoil, or a unique spectrun)}

\item{SpecATM_Sensor}{list. Includes diffuse and direct light
refractive index, specific absorption coefs and corresponding spectral bands}

\item{Parms2Estimate}{numeric. rank of variables to be inverted
to be estimated through inversion}

\item{input_prosail}{dataframe. full set of PROSAIL input variables}

\item{TypeLidf}{numeric. type of leaf inclination distribution function}

\item{prior_info}{list. prior mean, sd & weight of parms defined as xprior}

\item{Parms2Prior}{numeric. rank of parameters to be used with prior info
to modulate its importance}
}
\value{
fc estimates of the parameters
}
\description{
Merit function for PROSAIL inversion
}
