% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIS-data.R
\docType{data}
\name{MODIS}
\alias{MODIS}
\title{Spectral response function (SRF) for MODIS}
\format{
A list including
\itemize{
\item \emph{Spectral_Response}: the SRF of each spectral band of the sensor for a spectral domain compatible with PROSAIL
\item \emph{Spectral_Bands}: the spectral bands of the sensor (i.e. wavelength or name of the band as character)
\item \emph{OriginalBands}: original spectral bands for which SRF is defined (compatible with PROSAIL == 1 nm spectral sampling)
}
}
\usage{
MODIS
}
\description{
Spectral response function (SRF) for MODIS
}
\keyword{datasets}
