% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2char.R
\name{num2char}
\alias{num2char}
\title{downloads Sentinel-2 data and corresponding scene classification}
\usage{
num2char(val, nbdigits = 3)
}
\arguments{
\item{val}{numeric. value to be converted into string}

\item{nbdigits}{numeric. number pf digits for string}
}
\value{
valchar string corresponding to val with 0 if needed
}
\description{
downloads Sentinel-2 data and corresponding scene classification
}
