% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_nusvr.R
\name{optimize_nusvr}
\alias{optimize_nusvr}
\title{This function optimizes kernlab nu svr model using bayesian optimization}
\usage{
optimize_nusvr(lut, options = NULL)
}
\arguments{
\item{lut}{numeric. LUT of BRF used for training}

\item{options}{numeric. biophysical parameter corresponding to refl}
}
\value{
list of optimized hyperparameters
}
\description{
This function optimizes kernlab nu svr model using bayesian optimization
}
