% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pleiades_1B-data.R
\docType{data}
\name{Pleiades_1B}
\alias{Pleiades_1B}
\title{Spectral response function (SRF) for Pleiades-1B}
\format{
A list including
\itemize{
\item \emph{spectral_response}: the SRF of each spectral band of the sensor for a
spectral domain compatible with PROSAIL
\item \emph{spectral_bands}: the spectral bands of the sensor (i.e. wavelength or
name of the band as character)
\item \emph{original_bands}: original spectral bands for which SRF is defined
(compatible with PROSAIL == 1 nm spectral sampling)
\item \emph{central_wl}: Central wavelength (in nm) of the sensor spectral bands
}
}
\usage{
Pleiades_1B
}
\description{
Spectral response function (SRF) for Pleiades-1B
}
\keyword{datasets}
