% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sensor_simulation.R
\name{prepare_sensor_simulation}
\alias{prepare_sensor_simulation}
\title{This function converts high resolution spectral info into broader
spectral characteristics}
\usage{
prepare_sensor_simulation(spec_prospect, spec_soil, spec_atm, srf)
}
\arguments{
\item{spec_prospect}{list. optical constants required for PROSPECT}

\item{spec_soil}{list. dry soil and wet soil, or a unique soil sample if
psoil parameter is not inverted}

\item{spec_atm}{list. direct and diffuse radiation for clear conditions}

\item{srf}{list. Spectral response function, corresponding spectral bands,
and Original Bands}
}
\value{
spec_sensor list. list of input specs with sensor resolution
}
\description{
This function converts high resolution spectral info into broader
spectral characteristics
}
