% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_BRF.R, R/prosail-depcated.R
\name{Compute_BRF}
\alias{Compute_BRF}
\alias{prosail-deprecated}
\title{Deprecated functions in prosail}
\usage{
Compute_BRF(rdot, rsot, tts, SpecATM_Sensor, skyl = NULL)
}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
\item \code{\link{Compute_BRF}}: This function is deprecated, and will
be removed in the next version of this package.
Use \code{\link{compute_brf}} instead.
}
}
