% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prosail-package.R
\docType{package}
\name{prosail-package}
\alias{prosail-package}
\title{prosail: PROSAIL leaf and canopy radiative transfer model and inversion routines}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The package prosail includes the canopy radiative transfer model PROSAIL coupled with the leaf model PROSPECT-PRO, the latest version of the PROSPECT leaf model. It also includes inversion routines based on iterative optimization and hybrid method in order to estimate vegetation properties (leaf and canopy) from sensor measurements.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/jbferet/prosail}
  \item Report bugs at \url{https://gitlab.com/jbferet/prosail/issues}
}

}
\author{
\strong{Maintainer}: Jean-Baptiste Feret \email{jb.feret@teledetection.fr} (\href{https://orcid.org/0000-0002-0151-1334}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Jean-Baptiste Feret \email{jb.feret@teledetection.fr} (\href{https://orcid.org/0000-0002-0151-1334}{ORCID}) [copyright holder]
  \item Florian de Boissieu \email{florian.deboissieu@inrae.fr} (\href{https://orcid.org/0000-0002-2185-9952}{ORCID})
}

}
\keyword{internal}
