% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROSAIL_Hybrid_Apply.R
\name{PROSAIL_Hybrid_Apply}
\alias{PROSAIL_Hybrid_Apply}
\title{This function applies the regression models trained with PROSAIL_Hybrid_Train}
\usage{
PROSAIL_Hybrid_Apply(RegressionModels, Refl, progressBar = FALSE)
}
\arguments{
\item{RegressionModels}{list. List of regression models produced by PROSAIL_Hybrid_Train}

\item{Refl}{numeric. LUT of bidirectional reflectances factors used for training}

\item{progressBar}{boolean. should progressbar be displayed?}
}
\value{
HybridRes list. Estimated values corresponding to Refl. Includes
\itemize{
\item MeanEstimate = mean value for the ensemble regression model
\item StdEstimate = std value for the ensemble regression model
}
}
\description{
This function applies the regression models trained with PROSAIL_Hybrid_Train
}
