% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT.R
\name{PROSPECT}
\alias{PROSPECT}
\title{core function running PROSPECT}
\usage{
PROSPECT(
  SpecPROSPECT,
  Input_PROSPECT = NULL,
  N = 1.5,
  CHL = 40,
  CAR = 8,
  ANT = 0,
  BROWN = 0,
  EWT = 0.01,
  LMA = 0.008,
  PROT = 0,
  CBC = 0,
  alpha = 40
)
}
\arguments{
\item{SpecPROSPECT}{list. Includes optical constants
refractive index, specific absorption coefficients and corresponding spectral bands}

\item{Input_PROSPECT}{dataframe. full list of PROSPECT input parameters. Default = Set to NULL}

\item{N}{numeric. Leaf structure parameter}

\item{CHL}{numeric. Chlorophyll content (microg.cm-2)}

\item{CAR}{numeric. Carotenoid content (microg.cm-2)}

\item{ANT}{numeric. Anthocyain content (microg.cm-2)}

\item{BROWN}{numeric. Brown pigment content (Arbitrary units)}

\item{EWT}{numeric. Equivalent Water Thickness (g.cm-2)}

\item{LMA}{numeric. Leaf Mass per Area (g.cm-2)}

\item{PROT}{numeric. protein content  (g.cm-2)}

\item{CBC}{numeric. NonProtCarbon-based constituent content (g.cm-2)}

\item{alpha}{numeric. Solid angle for incident light at surface of leaf (simulation of roughness)}
}
\value{
leaf directional-hemisphrical reflectance and transmittance
}
\description{
core function running PROSPECT
}
