% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SpectralIndices.R
\name{readRasterBands}
\alias{readRasterBands}
\title{This function selects bands from a raster or stars object}
\usage{
readRasterBands(Refl, Bands, ReflFactor = 1)
}
\arguments{
\item{Refl}{RasterBrick, RasterStack or list. Raster bands in the order of SensorBands.}

\item{Bands}{numeric. rank of bands to be read in Refl}

\item{ReflFactor}{numeric. multiplying factor used to write reflectance in image (==10000 for S2)}
}
\value{
Robj list. R object (default is raster, stars if Refl is stars object)
}
\description{
This function selects bands from a raster or stars object
}
