% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sentinel_2-data.R
\docType{data}
\name{Sentinel_2}
\alias{Sentinel_2}
\title{Spectral response function (SRF) for Sentinel-2}
\format{
A list including
\itemize{
\item \emph{Spectral_Response}: the SRF of each spectral band of the sensor for a
spectral domain compatible with PROSAIL
\item \emph{Spectral_Bands}: the spectral bands of the sensor (i.e. wavelength or
name of the band as character)
\item \emph{Original_Bands}: original spectral bands for which SRF is defined
(compatible with PROSAIL == 1 nm spectral sampling)
\item \emph{Central_WL}: Central wavelength (in nm) of the sensor spectral bands
}
}
\usage{
Sentinel_2
}
\description{
Spectral response function (SRF) for Sentinel-2
}
\keyword{datasets}
