% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sentinel_2A-data.R
\docType{data}
\name{Sentinel_2A}
\alias{Sentinel_2A}
\title{Spectral response function (SRF) for Sentinel-2A}
\format{
A list including
\itemize{
\item \emph{Spectral_Response}: the SRF of each spectral band of the sensor for a spectral domain compatible with PROSAIL
\item \emph{Spectral_Bands}: the spectral bands of the sensor (i.e. wavelength or name of the band as character)
\item \emph{Original_Bands}: original spectral bands for which SRF is defined (compatible with PROSAIL == 1 nm spectral sampling)
\item \emph{Central_WL}: Central wavelength (in nm) of the sensor spectral bands
}
}
\usage{
Sentinel_2A
}
\description{
Spectral response function (SRF) for Sentinel-2A
}
\keyword{datasets}
