% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_compatibility_parms_to_estimate.R
\name{set_compatibility_parms_to_estimate}
\alias{set_compatibility_parms_to_estimate}
\title{make sure of the compatibility of variable names between versions of prosail}
\usage{
set_compatibility_parms_to_estimate(parms_to_estimate)
}
\arguments{
\item{parms_to_estimate}{list. PROSAIL input variables}
}
\value{
parms_to_estimate numeric. updated parms_to_estimate
}
\description{
make sure of the compatibility of variable names between versions of prosail
}
