% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_options_prosail.R
\name{set_options_prosail}
\alias{set_options_prosail}
\title{set options}
\usage{
set_options_prosail(fun, options = NULL)
}
\arguments{
\item{fun}{character. name of the function which has optional parameters}

\item{options}{list. including
\itemize{
\item cloudcover
\item path_S2tilinggrid
\item overwrite
\item geomAcq
\item additional_process
\item bands2correct
\item fraction_vegetation
\item RadiometricFilter
}}
}
\value{
options with default values when missing
}
\description{
set options
}
