% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilspect.R
\name{soilspect}
\alias{soilspect}
\title{Computation of soil optical properties with SOILSPECT
(derived from Hapke) for a given single scattering albedo spectrum:}
\usage{
soilspect(
  albss,
  rugo,
  thetav,
  thetas,
  phi,
  phase1,
  phase2,
  phase3 = NULL,
  phase4 = NULL
)
}
\arguments{
\item{albss}{= single scattering albedo}

\item{rugo}{= rugosity parameter}

\item{thetav}{= view zenith angle (radians)}

\item{thetas}{= solar azimut angle (radians)}

\item{phi}{= relative azimut angle (radians)}

\item{phase1}{= phase function parameter 1}

\item{phase2}{= phase function parameter 2}

\item{phase3}{= phase function parameter 3}

\item{phase4}{= phase function parameter 4}
}
\value{
bidir  = soil bidirectional reflectance
dirhem = soil directional hemispherical reflectance
hemhem = soil bihemispherical reflectance
}
\description{
Computation of soil optical properties with SOILSPECT
(derived from Hapke) for a given single scattering albedo spectrum:
}
