% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_atm-data.R
\docType{data}
\name{spec_atm}
\alias{spec_atm}
\title{direct and diffuse light for PROSAIL simulations}
\format{
A dataframe including 2101 rows and 10 columns
\itemize{
\item \emph{lambda}: wavelength, between 400 nm and 2500 nm
\item \emph{direct_light}: direct radiation solar spectra
\item \emph{diffuse_light}: diffuse radiation solar spectra
}
}
\usage{
spec_atm
}
\description{
corresponds to spectral bands, direct radiation solar spectra, diffuse
radiation solar spectra
}
\keyword{datasets}
