% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_prospect_full_range-data.R
\docType{data}
\name{spec_prospect_full_range}
\alias{spec_prospect_full_range}
\title{spec_prospect_full_range: optical constants defined for PROSPECT}
\format{
A dataframe including 2101 rows and 10 columns
\itemize{
\item \emph{lambda}: wavelength, between 400 nm and 2500 nm
\item \emph{nrefrac}: refractive index
\item \emph{sac_chl}: specific absorption coefficient for chlorophylls
\item \emph{sac_car}: specific absorption coefficient for carotenoids
\item \emph{sac_ant}: specific absorption coefficient for anthocyanins
\item \emph{sac_brown}: specific absorption coefficient for brown pigments
\item \emph{sac_ewt}: specific absorption coefficient for water content (Equivalent Water Thickness)
\item \emph{sac_lma}: specific absorption coefficient for dry matter (Leaf Mass per Area)
\item \emph{sac_prot}: specific absorption coefficient for proteins
\item \emph{sac_cbc}: specific absorption coefficient for carbon-base constituents
\item \emph{calctav_90}: transmissivity of a dielectric plane surface, averaged over
all directions of incidence & all polarizations for solid angle of 90 deg
\item \emph{calctav_40}: transmissivity of a dielectric plane surface, averaged over
all directions of incidence & all polarizations for solid angle of 40 deg
}
}
\usage{
spec_prospect_full_range
}
\description{
Corresponds to spectral bands, refractive index and specific absorption
coefficient for each chemica constituent, defined over the spectral domain
ranging from 400 nm to 2500 nm
}
\details{
The specific absorption coefficients were calibrated using experimental data.
The details of the calibration should be found in the following publications:
http://dx.doi.org/10.1016/j.rse.2017.03.004
https://doi.org/10.1016/j.rse.2020.112173
}
\keyword{datasets}
