% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_soil-data.R
\docType{data}
\name{spec_soil}
\alias{spec_soil}
\title{soil reflectance for PROSAIL simulations}
\format{
A dataframe including 2101 rows and 10 columns
\itemize{
\item \emph{lambda}: wavelength, between 400 nm and 2500 nm
\item \emph{max_refl}: minimum soil reflectance
\item \emph{min_refl}: maximum soil reflectance
}
}
\usage{
spec_soil
}
\description{
corresponds to spectral bands, maximum soil reflectance and minimum soil reflectance
}
\keyword{datasets}
