% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_soil_atbd_v2-data.R
\docType{data}
\name{spec_soil_atbd_v2}
\alias{spec_soil_atbd_v2}
\title{soil reflectance from Sentinel-2 ATBD v2 identified for PROSAIL simulations}
\format{
A dataframe including 2101 rows and 2 columns
\itemize{
\item \emph{lambda}: wavelength, between 400 nm and 2500 nm
\item \emph{soil_01} to \emph{soil_07}: soil reflectance samples
}
}
\usage{
spec_soil_atbd_v2
}
\description{
corresponds to spectral bands, and a selection of 7 soil reflectances
}
\keyword{datasets}
