% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_soil_ossl-data.R
\docType{data}
\name{spec_soil_ossl}
\alias{spec_soil_ossl}
\title{soil reflectance from OSSL database identified for PROSAIL simulations based
on S2 ATBD}
\format{
A dataframe including 2101 rows and 2 columns
\itemize{
\item \emph{lambda}: wavelength, between 400 nm and 2500 nm
\item \emph{soil_01} to \emph{soil_47}: soil reflectance samples
}
}
\usage{
spec_soil_ossl
}
\description{
corresponds to spectral bands, maximum soil reflectance and minimum soil reflectance
}
\keyword{datasets}
