% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecATM-data.R
\docType{data}
\name{SpecATM}
\alias{SpecATM}
\title{soil reflectance for PROSAIL simulations}
\format{
A dataframe including 2101 rows and 10 columns
\itemize{
\item \emph{lambda}: wavelength, between 400 nm and 2500 nm
\item \emph{Direct_Light}: direct radiation solar spectra
\item \emph{Diffuse_Light}: diffuse radiation solar spectra
}
}
\usage{
SpecATM
}
\description{
corresponds to spectral bands, dry soil and wet soil
}
\keyword{datasets}
