% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPROSPECT-data.R
\docType{data}
\name{SpecPROSPECT}
\alias{SpecPROSPECT}
\title{optical constants defined for PROSPECT}
\format{A dataframe including 2101 rows and 10 columns
\itemize{
\item \emph{lambda}: wavelength, between 400 nm and 2500 nm
\item \emph{nrefrac}: refractive index
\item \emph{SAC_CHL}: specific absorption coefficient for chlorophylls
\item \emph{SAC_CAR}: specific absorption coefficient for carotenoids
\item \emph{SAC_ANT}: specific absorption coefficient for anthocyanins
\item \emph{SAC_BROWN}: specific absorption coefficient for brown pigmnts
\item \emph{SAC_EWT}: specific absorption coefficient for EWT/water
\item \emph{SAC_LMA}: specific absorption coefficient for LMA/dry matter
\item \emph{SAC_PROT}: specific absorption coefficient for proteins
\item \emph{SAC_CBC}: specific absorption coefficient for carbon-base constituents
}}
\usage{
SpecPROSPECT
}
\description{
corresponds to spectral bands, refractive index and specific aborption coefficient
}
\keyword{datasets}
