% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPROSPECT_FullRange-data.R
\docType{data}
\name{SpecPROSPECT_FullRange}
\alias{SpecPROSPECT_FullRange}
\title{SpecPROSPECT_FullRange: optical constants defined for PROSPECT}
\format{
A dataframe including 2101 rows and 10 columns
\itemize{
\item \emph{lambda}: wavelength, between 400 nm and 2500 nm
\item \emph{nrefrac}: refractive index
\item \emph{SAC_CHL}: specific absorption coefficient for chlorophylls
\item \emph{SAC_CAR}: specific absorption coefficient for carotenoids
\item \emph{SAC_ANT}: specific absorption coefficient for anthocyanins
\item \emph{SAC_BROWN}: specific absorption coefficient for brown pigments
\item \emph{SAC_EWT}: specific absorption coefficient for water content (Equivalent Water Thickness)
\item \emph{SAC_LMA}: specific absorption coefficient for dry matter (Leaf Mass per Area)
\item \emph{SAC_PROT}: specific absorption coefficient for proteins
\item \emph{SAC_CBC}: specific absorption coefficient for carbon-base constituents
\item \emph{calctav_90}: transmissivity of a dielectric plane surface, averaged over
all directions of incidence & all polarizations for solid angle of 90 deg
\item \emph{calctav_40}: transmissivity of a dielectric plane surface, averaged over
all directions of incidence & all polarizations for solid angle of 40 deg
}
}
\usage{
SpecPROSPECT_FullRange
}
\description{
Corresponds to spectral bands, refractive index and specific absorption
coefficient for each chemica constituent, defined over the spectral domain
ranging from 400 nm to 2500 nm
}
\details{
The specific absorption coefficients were calibrated using experimental data.
The details of the calibration should be found in the following publications:
http://dx.doi.org/10.1016/j.rse.2017.03.004
https://doi.org/10.1016/j.rse.2020.112173
}
\keyword{datasets}
