% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecSOIL-data.R
\docType{data}
\name{SpecSOIL}
\alias{SpecSOIL}
\title{soil reflectance for PROSAIL simulations}
\format{
A dataframe including 2101 rows and 10 columns
\itemize{
\item \emph{lambda}: wavelength, between 400 nm and 2500 nm
\item \emph{Dry_Soil}: reflectance of dry soil
\item \emph{Wet_Soil}: reflectance of wet soil
}
}
\usage{
SpecSOIL
}
\description{
corresponds to spectral bands, dry soil and wet soil
}
\keyword{datasets}
