% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SENSOR.R
\name{SRF_Available}
\alias{SRF_Available}
\title{This function returns the list of al sensors for which SRF is available}
\usage{
SRF_Available()
}
\value{
list of sensors
}
\description{
This function returns the list of al sensors for which SRF is available
}
