% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volscatt.R
\name{volscatt}
\alias{volscatt}
\title{Compute volume scattering functions and interception coefficients
for given solar zenith, viewing zenith, azimuth and leaf inclination angle.}
\usage{
volscatt(tts, tto, psi, ttl)
}
\arguments{
\item{tts}{numeric. solar zenith}

\item{tto}{numeric. viewing zenith}

\item{psi}{numeric. azimuth}

\item{ttl}{numeric. leaf inclination angle}
}
\value{
res list. includes chi_s, chi_o, frho, ftau
}
\description{
Compute volume scattering functions and interception coefficients
for given solar zenith, viewing zenith, azimuth and leaf inclination angle.
}
