% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSPECT.R
\name{Warn_MissingInput}
\alias{Warn_MissingInput}
\title{Display warning message to inform about the parametrs default values
used during PROSPECT simulations}
\usage{
Warn_MissingInput(Input_PROSPECT, ExpectedParms)
}
\arguments{
\item{Input_PROSPECT}{list. Input parameters sent to PROSPECT by user}

\item{ExpectedParms}{list. Full set of parameters expected to run PROSPECT}
}
\value{
WhichMissing missing parameters
}
\description{
Display warning message to inform about the parametrs default values
used during PROSPECT simulations
}
