% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_parm_prior.R
\name{which_parm_prior}
\alias{which_parm_prior}
\title{function identifying for which parameters prior information is known}
\usage{
which_parm_prior(prior_mean, prior_sd)
}
\arguments{
\item{prior_mean}{list. mean value expected for a list of parameters}

\item{prior_sd}{list. standard deviation expected for a list
of parameters}
}
\value{
fc estimates of the parameters
}
\description{
function identifying for which parameters prior information is known
}
