% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_parms_to_invert.R
\name{which_parms_to_invert}
\alias{which_parms_to_invert}
\title{function identifying which parameters should be estimated during inversion}
\usage{
which_parms_to_invert(InitialGuess, LowerBound, UpperBound, ParmSet)
}
\arguments{
\item{InitialGuess}{list. Initial values during inversion}

\item{LowerBound}{list. Lower bound values during inversion}

\item{UpperBound}{list. Upper bound values during inversion}

\item{ParmSet}{list. Values for variables out of inversion}
}
\value{
res list. includes
\itemize{
\item Parms2Estimate
\item Parms2Set
\item InitialGuess
\item LowerBound
\item UpperBound
\item ParmSet
\item InVar
fc estimates of the parameters
}
}
\description{
function identifying which parameters should be estimated during inversion
}
