% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL_Inversion.R
\name{WhichParameters2Invert}
\alias{WhichParameters2Invert}
\title{function identifying which parameters should be estimated during inversion}
\usage{
WhichParameters2Invert(InitialGuess, LowerBound, UpperBound, ParmSet)
}
\arguments{
\item{InitialGuess}{list. Initial values during inversion}

\item{LowerBound}{list. Lower bound values during inversion}

\item{UpperBound}{list. Upper bound values during inversion}

\item{ParmSet}{list. Values for variables out of inversion}
}
\value{
fc estimates of the parameters
}
\description{
function identifying which parameters should be estimated during inversion
}
