% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL_Inversion.R
\name{WhichParmPrior}
\alias{WhichParmPrior}
\title{function identifying for which parameters prior information is known}
\usage{
WhichParmPrior(PriorInfoMean, PriorInfoSD)
}
\arguments{
\item{PriorInfoMean}{list. mean value expected for a list of parameters}

\item{PriorInfoSD}{list. dtandard deviation to the mean expected for a list of parameters}
}
\value{
fc estimates of the parameters
}
\description{
function identifying for which parameters prior information is known
}
