% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_prospect_to_sail.R
\name{adjust_prospect_to_sail}
\alias{adjust_prospect_to_sail}
\title{adjust prospect inputs and run prospect in preparation for 4SAIL or 4SAIL2}
\usage{
adjust_prospect_to_sail(
  sail_version,
  spec_sensor,
  input_prospect,
  chl,
  car,
  ant,
  brown,
  ewt,
  lma,
  prot,
  cbc,
  n_struct,
  alpha,
  fraction_brown,
  brown_lop = NULL
)
}
\arguments{
\item{sail_version}{character. should be '4SAIL' or '4SAIL2'}

\item{spec_sensor}{dataframe. spectral properties required to run PROSPECT}

\item{input_prospect}{dataframe. includes all prospect input parameters}

\item{chl}{numeric. Chlorophyll content (microg.cm-2)}

\item{car}{numeric. Carotenoid content (microg.cm-2)}

\item{ant}{numeric. Anthocyain content (microg.cm-2)}

\item{brown}{numeric. Brown pigment content (Arbitrary units)}

\item{ewt}{numeric. Equivalent Water Thickness (g.cm-2)}

\item{lma}{numeric. Leaf Mass per Area (g.cm-2)}

\item{prot}{numeric. protein content  (g.cm-2)}

\item{cbc}{numeric. NonprotCarbon-based constituent content (g.cm-2)}

\item{n_struct}{numeric. Leaf structure parameter}

\item{alpha}{numeric. Solid angle for incident light at surface of leaf
(simulation of roughness)}

\item{fraction_brown}{numeric. fraction of brown vegetation (0-1)}

\item{brown_lop}{dataframe. brown leaf optical properties, when available}
}
\value{
invisible
}
\description{
adjust prospect inputs and run prospect in preparation for 4SAIL or 4SAIL2
}
