% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL_HybridInversion.R
\name{apply_noise_AddMult}
\alias{apply_noise_AddMult}
\title{This function applies additive and multiplicative noise to BRF data
\itemize{
\item if AdditiveNoise and MultiplicativeNoise are defined with a unique value,
noise is homogeneous across all spectrum
\item if AdditiveNoise and MultiplicativeNoise are the same length as the number
of spectral bands (rows in BRF_LUT), noise is specific to each spectral band
}}
\usage{
apply_noise_AddMult(BRF_LUT, AdditiveNoise = 0.01, MultiplicativeNoise = 0.02)
}
\arguments{
\item{BRF_LUT}{numeric. BRF LUT}

\item{AdditiveNoise}{numeric. additive noise (0 = 0\%, 1 = 100\%)}

\item{MultiplicativeNoise}{numeric. multiplicative noise (0 = 0\%, 1 = 100\%)}
}
\value{
BRF_LUT_Noise numeric.
}
\description{
This function applies additive and multiplicative noise to BRF data
\itemize{
\item if AdditiveNoise and MultiplicativeNoise are defined with a unique value,
noise is homogeneous across all spectrum
\item if AdditiveNoise and MultiplicativeNoise are the same length as the number
of spectral bands (rows in BRF_LUT), noise is specific to each spectral band
}
}
