% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_noise_AddMult.R
\name{apply_noise_AddMult}
\alias{apply_noise_AddMult}
\title{This function applies additive and multiplicative noise to BRF data
\itemize{
\item if additive_noise and multiplicative_noise are defined with a unique value,
noise is homogeneous across all spectrum
\item if additive_noise and multiplicative_noise are the same length as the number
of spectral bands (rows in BRF_LUT), noise is specific to each spectral band
}}
\usage{
apply_noise_AddMult(
  BRF_LUT,
  additive_noise = 0.01,
  multiplicative_noise = 0.02
)
}
\arguments{
\item{BRF_LUT}{numeric. BRF LUT}

\item{additive_noise}{numeric. additive noise (0 = 0\%, 1 = 100\%)}

\item{multiplicative_noise}{numeric. multiplicative noise (0 = 0\%, 1 = 100\%)}
}
\value{
BRF_LUT_Noise numeric.
}
\description{
This function applies additive and multiplicative noise to BRF data
\itemize{
\item if additive_noise and multiplicative_noise are defined with a unique value,
noise is homogeneous across all spectrum
\item if additive_noise and multiplicative_noise are the same length as the number
of spectral bands (rows in BRF_LUT), noise is specific to each spectral band
}
}
