% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_noise_atbd.R
\name{apply_noise_atbd}
\alias{apply_noise_atbd}
\title{This function applies noise defined in S2 ATBD to reflectance
look up table}
\usage{
apply_noise_atbd(refl_lut)
}
\arguments{
\item{refl_lut}{numeric. reflectance look up table}
}
\value{
refl_lut_noise numeric.
}
\description{
This function applies noise defined in S2 ATBD to reflectance
look up table
}
