% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_noise_atbd.R
\name{apply_noise_atbd}
\alias{apply_noise_atbd}
\title{This function applies noise defined in S2 ATBD to BRF LUT}
\usage{
apply_noise_atbd(BRF_LUT)
}
\arguments{
\item{BRF_LUT}{numeric. BRF LUT}
}
\value{
BRF_LUT_Noise numeric.
}
\description{
This function applies noise defined in S2 ATBD to BRF LUT
}
