% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_noise_atbd.R
\name{apply_noise_atbd}
\alias{apply_noise_atbd}
\title{This function applies noise defined in S2 ATBD to BRF LUT}
\usage{
apply_noise_atbd(brf_lut)
}
\arguments{
\item{brf_lut}{numeric. BRF LUT}
}
\value{
brf_lut_noise numeric.
}
\description{
This function applies noise defined in S2 ATBD to BRF LUT
}
