% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL_LUT.R
\name{Apply_Noise_LUT}
\alias{Apply_Noise_LUT}
\title{This function applied noise on a matrix}
\usage{
Apply_Noise_LUT(LUT, NoiseLevel, NoiseType = "relative")
}
\arguments{
\item{LUT}{numeric. Matrix including data to add noise to}

\item{NoiseLevel}{numeric. value of the normal noise proportional to LUT to apply on LUT}

\item{NoiseType}{character.
\itemize{
\item relative: noise proportional to actual value to add noise to
\item absolute: noise not proportional to actual value to add noise to
}}
}
\value{
LUT_Noise numeric. Matrix including data with added noise
}
\description{
This function applied noise on a matrix
}
