% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_noise_lut.R
\name{apply_noise_lut}
\alias{apply_noise_lut}
\title{This function applied noise on a matrix}
\usage{
apply_noise_lut(LUT, NoiseLevel, NoiseType = "relative")
}
\arguments{
\item{LUT}{numeric. Matrix including data to add noise to}

\item{NoiseLevel}{numeric. value of the normal noise proportional to LUT to
apply on LUT}

\item{NoiseType}{character.
\itemize{
\item relative: noise proportional to actual value to add noise to
\item absolute: noise not proportional to actual value to add noise to
}}
}
\value{
LUT_Noise numeric. Matrix including data with added noise
}
\description{
This function applied noise on a matrix
}
