% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_noise_lut.R
\name{apply_noise_lut}
\alias{apply_noise_lut}
\title{This function applied noise on a matrix}
\usage{
apply_noise_lut(lut, noise_level = 0, noise_type = "relative")
}
\arguments{
\item{lut}{numeric. Matrix including data to add noise to}

\item{noise_level}{numeric. value of the normal noise proportional to lut to
apply on lut}

\item{noise_type}{character.
\itemize{
\item relative: noise proportional to actual value to add noise to
\item absolute: noise not proportional to actual value to add noise to
}}
}
\value{
lut_noise numeric. Matrix including data with added noise
}
\description{
This function applied noise on a matrix
}
