% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_prosail_inversion_tiles.R
\name{apply_prosail_inversion_tiles}
\alias{apply_prosail_inversion_tiles}
\title{This function applies SVR model on raster data in order to estimate
vegetation biophysical properties}
\usage{
apply_prosail_inversion_tiles(
  raster_path,
  hybrid_model,
  output_dir,
  selected_bands,
  band_names,
  mask_path = NULL,
  multiplying_factor = 10000,
  tile_size = 10000,
  filetype = "GTiff"
)
}
\arguments{
\item{raster_path}{character. path for a raster file}

\item{hybrid_model}{list. hybrid models produced from train_prosail_inversion
each element of the list corresponds to a set of hybrid models for a
vegetation parameter}

\item{output_dir}{character. path for directory where results are written}

\item{selected_bands}{list. list of spectral bands to be selected from raster
(identified by name of vegetation parameter)}

\item{band_names}{character. spectral bands corresponding to the raster}

\item{mask_path}{character. path for binary mask defining ON (1) and OFF (0)
pixels in the raster}

\item{multiplying_factor}{numeric. multiplying factor used to write
reflectance in the raster
--> PROSAIL simulates reflectance between 0 and 1, and raster data expected
in the same range}

\item{tile_size}{numeric. size of individual tiles in meters}

\item{filetype}{character. driver for raster file}
}
\value{
res character. path for output files corresponding to
biophysical properties
}
\description{
This function applies SVR model on raster data in order to estimate
vegetation biophysical properties
}
