% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_sensor_characteristics.R
\name{apply_sensor_characteristics}
\alias{apply_sensor_characteristics}
\title{reads spectral response from known sensor
spectral response from Sentinel-2 is already defined}
\usage{
apply_sensor_characteristics(wvl, refl, srf)
}
\arguments{
\item{wvl}{numeric. spectral sampling of the input spectral data}

\item{refl}{numeric. input spectral data (unique sample)}

\item{srf}{list. Spectral response & Spectral Bands of the sensor & Original
Bands for which srf is defined}
}
\value{
refl_sensor numeric. Output spectral data, sensor resolution
}
\description{
reads spectral response from known sensor
spectral response from Sentinel-2 is already defined
}
