% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spectral_sampling.R
\name{check_spectral_sampling}
\alias{check_spectral_sampling}
\title{check if spectral sampling is identical between spec_prospect, spec_soil,
spec_atm}
\usage{
check_spectral_sampling(spec_prospect, spec_soil, spec_atm)
}
\arguments{
\item{spec_prospect}{list. Includes optical constants required for PROSPECT}

\item{spec_soil}{list. Includes either dry soil and wet soil, or a unique
soil sample if the psoil parameter is not inverted}

\item{spec_atm}{list. direct and diffuse radiation for clear conditions}
}
\value{
invisible
}
\description{
check if spectral sampling is identical between spec_prospect, spec_soil,
spec_atm
}
