% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_BRF.R
\name{compute_BRF}
\alias{compute_BRF}
\title{Computes BRF based on outputs from PROSAIL & SZA}
\usage{
compute_BRF(rdot, rsot, tts, SpecATM_Sensor, skyl = NULL)
}
\arguments{
\item{rdot}{numeric. Hemispherical-directional refl factor viewing direction}

\item{rsot}{numeric. Bi-directional reflectance factor}

\item{tts}{numeric. Solar zenith angle}

\item{SpecATM_Sensor}{list. direct and diffuse radiation for clear conditions}

\item{skyl}{numeric. values for skyl}
}
\value{
BRF numeric. Bidirectional reflectance factor
}
\description{
The direct and diffuse light are taken into account as proposed by:
Francois et al. (2002) Conversion of 400-1100 nm vegetation albedo
measurements into total shortwave broadband albedo using a canopy
radiative transfer model, Agronomie
Es = direct
Ed = diffuse
}
