% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL.R
\name{Compute_BRF}
\alias{Compute_BRF}
\title{Computes bidirectional reflectance factor based on outputs from PROSAIL and sun position
The direct and diffuse light are taken into account as proposed by:
Francois et al. (2002) Conversion of 400-1100 nm vegetation albedo
measurements into total shortwave broadband albedo using a canopy
radiative transfer model, Agronomie
Es = direct
Ed = diffuse}
\usage{
Compute_BRF(rdot, rsot, tts, SpecATM_Sensor)
}
\arguments{
\item{rdot}{numeric. Hemispherical-directional reflectance factor in viewing direction}

\item{rsot}{numeric. Bi-directional reflectance factor}

\item{tts}{numeric. Solar zenith angle}

\item{SpecATM_Sensor}{list. direct and diffuse light for clear conditions}
}
\value{
BRF numeric. Bidirectional reflectance factor
}
\description{
Computes bidirectional reflectance factor based on outputs from PROSAIL and sun position
The direct and diffuse light are taken into account as proposed by:
Francois et al. (2002) Conversion of 400-1100 nm vegetation albedo
measurements into total shortwave broadband albedo using a canopy
radiative transfer model, Agronomie
Es = direct
Ed = diffuse
}
