% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fapar.R
\name{compute_fapar}
\alias{compute_fapar}
\title{Computes fraction of absorbed photosyntehtically active radiation fAPAR}
\usage{
compute_fapar(abs_dir, abs_hem, tts, spec_atm_sensor, par_range = c(400, 700))
}
\arguments{
\item{abs_dir}{numeric. fraction of direct light absorbed}

\item{abs_hem}{numeric. fraction of diffuse light absorbed}

\item{tts}{numeric. Solar zenith angle}

\item{spec_atm_sensor}{list. direct & diffuse radiation for clear conditions}

\item{par_range}{numeric. range (in nm) of spectral domain to integrate for
computation of fAPAR}
}
\value{
fapar numeric. fAPAR
}
\description{
The direct and diffuse light are taken into account as proposed by:
Francois et al. (2002) Conversion of 400-1100 nm vegetation albedo
measurements into total shortwave broadband albedo using a canopy
radiative transfer model, Agronomie
Es = direct
Ed = diffuse
}
