% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SENSOR.R
\name{Compute_SRF}
\alias{Compute_SRF}
\title{Computes spectral response function based on wavelength and FWHM characteristics}
\usage{
Compute_SRF(wvl, FWHM)
}
\arguments{
\item{wvl}{numeric. spectral sampling of the sensor}

\item{FWHM}{numeric. Full Width Half Maximum for each spectral band}
}
\value{
SRF list. Information about spectral response Spectral Bands of the sensor and Original Bands for which SRF is defined
}
\description{
Computes spectral response function based on wavelength and FWHM characteristics
}
