% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_SRF.R
\name{compute_SRF}
\alias{compute_SRF}
\title{Computes spectral response function based on wavelength and FWHM
characteristics}
\usage{
compute_SRF(wvl, FWHM, SensorName = "Custom")
}
\arguments{
\item{wvl}{numeric. spectral sampling of the sensor}

\item{FWHM}{numeric. Full Width Half Maximum for each spectral band}

\item{SensorName}{character. sensor name}
}
\value{
SRF list. spectral response, sensor Spectral Bands & Original Bands
for which SRF is defined
}
\description{
Computes spectral response function based on wavelength and FWHM
characteristics
}
