% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL_Inversion.R
\name{CostVal_RMSE_PROSAIL}
\alias{CostVal_RMSE_PROSAIL}
\title{Value of the cost criterion to minimize during PROSAIL inversion}
\usage{
CostVal_RMSE_PROSAIL(brfMES, brfMOD)
}
\arguments{
\item{brfMES}{numeric. Measured bidirectional reflectance}

\item{brfMOD}{numeric. Simulated bidirectional reflectance}
}
\value{
res list. Includes Parms2Estimate, Parms2Set,
InitialGuess, LowerBound, UpperBound, ParmSet, InVar
}
\description{
Value of the cost criterion to minimize during PROSAIL inversion
}
