% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL.R
\name{fourSAIL}
\alias{fourSAIL}
\title{Performs PROSAIL simulation based on a set of combinations of input parameters}
\usage{
fourSAIL(
  LeafOptics,
  TypeLidf = 2,
  LIDFa = NULL,
  LIDFb = NULL,
  lai = NULL,
  q = NULL,
  tts = NULL,
  tto = NULL,
  psi = NULL,
  rsoil = NULL
)
}
\arguments{
\item{LeafOptics}{list. Includes leaf optical properties (reflectance and transmittance)
and corresponding spectral bands}

\item{TypeLidf}{numeric. Type of leaf inclination distribution function}

\item{LIDFa}{numeric.
if TypeLidf ==1, controls the average leaf slope
if TypeLidf ==2, controls the average leaf angle}

\item{LIDFb}{numeric.
if TypeLidf ==1, controls the distribution's bimodality
if TypeLidf ==2, unused}

\item{lai}{numeric. Leaf Area Index}

\item{q}{numeric. Hot Spot parameter}

\item{tts}{numeric. Sun zeith angle}

\item{tto}{numeric. Observer zeith angle}

\item{psi}{numeric. Azimuth Sun / Observer}

\item{rsoil}{numeric. Soil reflectance}
}
\value{
list. rdot,rsot,rddt,rsdt
rdot: hemispherical-directional reflectance factor in viewing direction
rsot: bi-directional reflectance factor
rsdt: directional-hemispherical reflectance factor for solar incident flux
rddt: bi-hemispherical reflectance factor
}
\description{
Performs PROSAIL simulation based on a set of combinations of input parameters
}
