% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourSAIL2.R
\name{fourSAIL2}
\alias{fourSAIL2}
\title{4SAIL2 simulation based on a set of combinations of input parameters}
\usage{
fourSAIL2(
  leaf_green,
  leaf_brown,
  type_lidf = 2,
  lidf_a = 60,
  lidf_b = NULL,
  lai = 3,
  hotspot = 0.1,
  tts = 30,
  tto = 0,
  psi = 60,
  rsoil = NULL,
  fraction_brown = 0.5,
  diss = 0.5,
  cv = 1,
  zeta = 1
)
}
\arguments{
\item{leaf_green}{dataframe. leaf optical properties #1 (e.g. green veg)}

\item{leaf_brown}{dataframe. leaf optical properties #2 (e.g. brown veg)}

\item{type_lidf}{numeric. Type of leaf inclination distribution function}

\item{lidf_a}{numeric.
if type_lidf ==1, controls the average leaf slope
if type_lidf ==2, controls the average leaf angle}

\item{lidf_b}{numeric.
if type_lidf ==1, unused
if type_lidf ==2, controls the distribution's bimodality}

\item{lai}{numeric. Leaf Area Index}

\item{hotspot}{numeric. Hot Spot parameter = ratio of the correlation length of
leaf projections in the horizontal plane and the canopy height
(doi:10.1016/j.rse.2006.12.013)}

\item{tts}{numeric. Sun zeith angle}

\item{tto}{numeric. Observer zeith angle}

\item{psi}{numeric. Azimuth Sun / Observer}

\item{rsoil}{numeric. Soil reflectance}

\item{fraction_brown}{numeric. Fraction of brown leaf area}

\item{diss}{numeric. Layer dissociation factor}

\item{cv}{numeric. vertical crown cover percentage
= \% ground area covered with crowns as seen from nadir direction}

\item{zeta}{numeric. Tree shape factor
= ratio of crown diameter to crown height}
}
\value{
list. rdot,rsot,rddt,rsdt
rdot: hemispherical-directional reflectance factor in viewing direction
rsot: bi-directional reflectance factor
rsdt: directional-hemispherical reflectance factor for solar incident flux
rddt: bi-hemispherical reflectance factor
fcover: Fraction of green Vegetation Cover (= 1 - beam transmittance in the
target-view path)
abs_dir: canopy absorptance for direct solar incident flux
abs_hem: canopy absorptance for hemispherical diffuse incident flux
rsdstar: contribution of direct solar incident flux to albedo
rddstar: contribution of hemispherical diffuse incident flux to albedo
}
\description{
4SAIL2 simulation based on a set of combinations of input parameters
}
