% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL.R
\name{fourSAIL2}
\alias{fourSAIL2}
\title{Performs PRO4SAIL2 simulation based on a set of combinations of input parameters}
\usage{
fourSAIL2(
  leafgreen,
  leafbrown,
  TypeLidf = 2,
  LIDFa = 60,
  LIDFb = NULL,
  lai = 3,
  q = 0.1,
  tts = 30,
  tto = 0,
  psi = 60,
  rsoil = NULL,
  fraction_brown = 0.5,
  diss = 0.5,
  Cv = 1,
  Zeta = 1
)
}
\arguments{
\item{leafgreen}{dataframe. leaf optical properties #1 (e.g. green vegetation)}

\item{leafbrown}{dataframe. leaf optical properties #2 (e.g. brown vegetation)}

\item{TypeLidf}{numeric. Type of leaf inclination distribution function}

\item{LIDFa}{numeric.
if TypeLidf ==1, controls the average leaf slope
if TypeLidf ==2, controls the average leaf angle}

\item{LIDFb}{numeric.
if TypeLidf ==1, unused
if TypeLidf ==2, controls the distribution's bimodality}

\item{lai}{numeric. Leaf Area Index}

\item{q}{numeric. Hot Spot parameter = ratio of the correlation length of
leaf projections in the horizontal plane and the canopy height
(doi:10.1016/j.rse.2006.12.013)}

\item{tts}{numeric. Sun zeith angle}

\item{tto}{numeric. Observer zeith angle}

\item{psi}{numeric. Azimuth Sun / Observer}

\item{rsoil}{numeric. Soil reflectance}

\item{fraction_brown}{numeric. Fraction of brown leaf area}

\item{diss}{numeric. Layer dissociation factor}

\item{Cv}{numeric. vertical crown cover percentage
= \% ground area covered with crowns as seen from nadir direction}

\item{Zeta}{numeric. Tree shape factor
= ratio of crown diameter to crown height}
}
\value{
list. rdot,rsot,rddt,rsdt
rdot: hemispherical-directional reflectance factor in viewing direction
rsot: bi-directional reflectance factor
rsdt: directional-hemispherical reflectance factor for solar incident flux
rddt: bi-hemispherical reflectance factor
fcover: Fraction of green Vegetation Cover (= 1 - beam transmittance in the target-view path)
abs_dir: canopy absorptance for direct solar incident flux
abs_hem: canopy absorptance for hemispherical diffuse incident flux
rsdstar: contribution of direct solar incident flux to albedo
rddstar: contribution of hemispherical diffuse incident flux to albedo
}
\description{
Performs PRO4SAIL2 simulation based on a set of combinations of input parameters
}
