% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_LUT_4SAIL.R
\name{generate_LUT_4SAIL}
\alias{generate_LUT_4SAIL}
\title{generates a LUT of 4SAIL outputs based on a table of input variables for PRO4SAIL model}
\usage{
generate_LUT_4SAIL(
  input_prosail,
  SpecPROSPECT,
  SpecSOIL,
  SpecATM,
  band_names = NULL,
  SAILversion = "4SAIL",
  brown_lop = NULL
)
}
\arguments{
\item{input_prosail}{list. PROSAIL input variables}

\item{SpecPROSPECT}{list. Includes optical constants required for PROSPECT}

\item{SpecSOIL}{list. Includes either dry soil and wet soil, or a unique
soil sample if the psoil parameter is not inverted}

\item{SpecATM}{list. direct and diffuse radiation for clear conditions}

\item{band_names}{character. Name of the spectral bands of the sensor}

\item{SAILversion}{character. choose between 4SAIL and 4SAIL2}

\item{brown_lop}{list. Defines optical properties for brown vegetation
\itemize{
\item WVL, Reflectance, Transmittance
\item Set to NULL if use PROSPECT to generate it
}}
}
\value{
4SAIL_LUT numeric. matrix of 4SAIL outputs (rdot, rsot, rsdt, rddt)
corresponding to input_prosail
}
\description{
generates a LUT of 4SAIL outputs based on a table of input variables for PRO4SAIL model
}
