% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_LUT_BRF.R
\name{generate_lut_brf}
\alias{generate_lut_brf}
\title{generates a LUT of brf based on a table of input variables for prosail model}
\usage{
generate_lut_brf(
  input_prosail,
  spec_prospect,
  spec_soil,
  spec_atm,
  band_names = NULL,
  SAILversion = "4SAIL",
  brown_lop = NULL
)
}
\arguments{
\item{input_prosail}{list. PROSAIL input variables}

\item{spec_prospect}{list. Includes optical constants required for PROSPECT}

\item{spec_soil}{list. Includes either a set of OSSL library reflectance
spectra, or minimum reflectance and maximum reflectance}

\item{spec_atm}{list. Includes direct and diffuse radiation for
clear conditions}

\item{band_names}{character. Name of the spectral bands of the sensor}

\item{SAILversion}{character. choose between 4SAIL and 4SAIL2}

\item{brown_lop}{list. Defines optical properties for brown vegetation
\itemize{
\item WVL, Reflectance, Transmittance
\item Set to NULL if use PROSPECT to generate it
}}
}
\value{
brf numeric. matrix of brf corresponding to input_prosail
}
\description{
generates a LUT of brf based on a table of input variables for prosail model
}
