% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_LUT_PROSAIL.R
\name{generate_LUT_PROSAIL}
\alias{generate_LUT_PROSAIL}
\title{This function generates a LUT of PROSAIL outputs, including BRF, fAPAR,
fCover and albedo, based on a table of input variables for PRO4SAIL model}
\usage{
generate_LUT_PROSAIL(
  input_prosail,
  SpecPROSPECT,
  SpecSOIL,
  SpecATM,
  band_names = NULL,
  SAILversion = "4SAIL",
  brown_lop = NULL,
  progress = TRUE
)
}
\arguments{
\item{input_prosail}{list. PROSAIL input variables}

\item{SpecPROSPECT}{list. Includes optical constants required for PROSPECT}

\item{SpecSOIL}{list. Includes either dry soil and wet soil, or a unique
soil sample if the psoil parameter is not inverted}

\item{SpecATM}{list. direct and diffuse radiation for clear conditions}

\item{band_names}{character. Name of the spectral bands of the sensor}

\item{SAILversion}{character. choose between 4SAIL and 4SAIL2}

\item{brown_lop}{list. Defines optical properties for brown vegetation
\itemize{
\item WVL, Reflectance, Transmittance
\item Set to NULL if use PROSPECT to generate it
}}

\item{progress}{boolean. set TRUE for progress bar during production of LUT}
}
\value{
LUT numeric. list of BRF, fCover, fAPAR and albedo corresponding
to input_prosail
}
\description{
This function generates a LUT of PROSAIL outputs, including BRF, fAPAR,
fCover and albedo, based on a table of input variables for PRO4SAIL model
}
