% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_LUT_PROSAIL.R
\name{generate_LUT_PROSAIL}
\alias{generate_LUT_PROSAIL}
\title{This function generates a LUT of PROSAIL outputs, including BRF, fAPAR,
fCover and albedo, based on a table of input variables for PRO4SAIL model}
\usage{
generate_LUT_PROSAIL(
  InputPROSAIL,
  SpecPROSPECT,
  SpecSOIL,
  SpecATM,
  BandNames = NULL,
  SAILversion = "4SAIL",
  BrownLOP = NULL,
  progress = TRUE
)
}
\arguments{
\item{InputPROSAIL}{list. PROSAIL input variables}

\item{SpecPROSPECT}{list. Includes optical constants required for PROSPECT}

\item{SpecSOIL}{list. Includes either dry soil and wet soil, or a unique soil sample if the psoil parameter is not inverted}

\item{SpecATM}{list. Includes direct and diffuse radiation for clear conditions}

\item{BandNames}{character. Name of the spectral bands of the sensor}

\item{SAILversion}{character. choose between 4SAIL and 4SAIL2}

\item{BrownLOP}{list. Defines optical properties for brown vegetation, if not NULL
\itemize{
\item WVL, Reflectance, Transmittance
\item Set to NULL if use PROSPECT to generate it
}}

\item{progress}{boolean. set TRUE to get progress bar during production of LUT}
}
\value{
LUT numeric. list of BRF, fCover, fAPAR and albedo corresponding to InputPROSAIL
}
\description{
This function generates a LUT of PROSAIL outputs, including BRF, fAPAR,
fCover and albedo, based on a table of input variables for PRO4SAIL model
}
