% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_lut_prosail.R
\name{generate_lut_prosail}
\alias{generate_lut_prosail}
\title{This function generates a LUT of PROSAIL outputs, including brf, fapar,
fcover and albedo, based on a table of input variables for prosail model}
\usage{
generate_lut_prosail(
  input_prosail,
  spec_prospect,
  spec_soil,
  spec_atm,
  band_names = NULL,
  SAILversion = "4SAIL",
  brown_lop = NULL,
  progress = TRUE
)
}
\arguments{
\item{input_prosail}{list. PROSAIL input variables}

\item{spec_prospect}{list. Includes optical constants required for PROSPECT}

\item{spec_soil}{list. Includes either a set of OSSL library reflectance
spectra, or minimum reflectance and maximum reflectance}

\item{spec_atm}{list. direct and diffuse radiation for clear conditions}

\item{band_names}{character. Name of the spectral bands of the sensor}

\item{SAILversion}{character. choose between 4SAIL and 4SAIL2}

\item{brown_lop}{list. Defines optical properties for brown vegetation
\itemize{
\item WVL, Reflectance, Transmittance
\item Set to NULL if use PROSPECT to generate it
}}

\item{progress}{boolean. set TRUE for progress bar during production of LUT}
}
\value{
LUT numeric. list of brf, fcover, fapar and albedo corresponding
to input_prosail
}
\description{
This function generates a LUT of PROSAIL outputs, including brf, fapar,
fcover and albedo, based on a table of input variables for prosail model
}
