% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL_LUT.R
\name{get_codistributions}
\alias{get_codistributions}
\title{This function adjusts variable values based on co-distribution rules as defined in ATBD
co-distributions are all related to LAI}
\usage{
get_codistributions(V, LAI, MaxLAI, Vmin0, Vmax0, VminLAImax, VmaxLAImax)
}
\arguments{
\item{V}{numeric.}

\item{LAI}{numeric.}

\item{MaxLAI}{numeric.}

\item{Vmin0}{numeric.}

\item{Vmax0}{numeric.}

\item{VminLAImax}{numeric.}

\item{VmaxLAImax}{numeric.}
}
\value{
Vstar numeric.
}
\description{
This function adjusts variable values based on co-distribution rules as defined in ATBD
co-distributions are all related to LAI
}
