% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_LUT_input.R
\name{get_default_lut_input}
\alias{get_default_lut_input}
\title{Sets default values for PROSAIL LUT simulation when not defined by user}
\usage{
get_default_lut_input(
  type_distrib = NULL,
  gaussian_distrib = NULL,
  minval = NULL,
  maxval = NULL
)
}
\arguments{
\item{type_distrib}{list. uniform/Gaussian distribution to apply.}

\item{gaussian_distrib}{list. mean & sd for parms with gaussian dist}

\item{minval}{list. minimum value for a list of parameters randomly produced}

\item{maxval}{list. maximum value for a list of parameters randomly produced}
}
\value{
res list. default values corresponding to NULL input parameters
}
\description{
Sets default values for PROSAIL LUT simulation when not defined by user
}
