% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_LUT_input.R
\name{get_default_LUT_input}
\alias{get_default_LUT_input}
\title{Sets default values for PROSAIL LUT simulation when not defined by user}
\usage{
get_default_LUT_input(
  TypeDistrib = NULL,
  GaussianDistrib = NULL,
  minval = NULL,
  maxval = NULL
)
}
\arguments{
\item{TypeDistrib}{list. uniform/Gaussian distribution to apply.}

\item{GaussianDistrib}{list. Mean & STD for parms with gaussian distribution}

\item{minval}{list. minimum value for a list of parameters randomly produced}

\item{maxval}{list. maximum value for a list of parameters randomly produced}
}
\value{
res list. default values corresponding to NULL input parameters
}
\description{
Sets default values for PROSAIL LUT simulation when not defined by user
}
