% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distribution_input_prosail.R
\name{get_distribution_input_prosail}
\alias{get_distribution_input_prosail}
\title{This function generates distribution of biophysical parameters used as input
parameters in PRO4SAIL}
\usage{
get_distribution_input_prosail(
  minval = NULL,
  maxval = NULL,
  parm_set = NULL,
  nb_samples = 2000,
  TypeDistrib = NULL,
  Mean = NULL,
  Std = NULL
)
}
\arguments{
\item{minval}{list. minimum value for a list of parameters randomly produced}

\item{maxval}{list. maximum value for a list of parameters randomly produced}

\item{parm_set}{list. parameters to be set to a given value}

\item{nb_samples}{numeric. Number of samples to be generated}

\item{TypeDistrib}{list. uniform/Gaussian distribution to be applied.}

\item{Mean}{list. mean value for parameters with Gaussian distribution}

\item{Std}{list. standard deviation for parameters with Gaussian distribution}
}
\value{
input_prosail list. list of nb_samples input parameters for
PRO4SAIL simulation
}
\description{
This function generates distribution of biophysical parameters used as input
parameters in PRO4SAIL
}
