% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_PROSAIL.R
\name{get_input_PROSAIL}
\alias{get_input_PROSAIL}
\title{generate InputPROSAIL, following
\itemize{
\item either distribution defined in ATBD
\item or distribution defined in user
}}
\usage{
get_input_PROSAIL(
  atbd = FALSE,
  GeomAcq = NULL,
  Codist_LAI = TRUE,
  minval = NULL,
  maxval = NULL,
  TypeDistrib = NULL,
  GaussianDistrib = NULL,
  ParmSet = NULL,
  nbSamples = 2000,
  verbose = FALSE
)
}
\arguments{
\item{atbd}{boolean. TRUE to apply input parameter distribution from ATBD}

\item{GeomAcq}{list. geom of acquisiton: min and max val for tts, tto & psi}

\item{Codist_LAI}{boolean. set TYRUE if codistribution with LAI accounted for}

\item{minval}{list. min val for input parameters}

\item{maxval}{list. max val for input parameters}

\item{TypeDistrib}{list. Type of distribution: 'Uniform' or 'Gaussian'}

\item{GaussianDistrib}{list. Mean & SD for parms sampled with gaussian dist}

\item{ParmSet}{list. list of input parameters set to a specific value}

\item{nbSamples}{numeric. number of samples in training LUT}

\item{verbose}{boolean. when set to TRUE, prints message if hyperparameter
adjustment performed during training}
}
\value{
InputPROSAIL
}
\description{
generate InputPROSAIL, following
\itemize{
\item either distribution defined in ATBD
\item or distribution defined in user
}
}
