% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_s2_geometry_from_SAFE.R
\name{get_s2_geometry_from_SAFE}
\alias{get_s2_geometry_from_SAFE}
\title{This function returns geometry of acquisition for S2 image processed with
Sen2Cor
\itemize{
\item SZA = list of sun zenith angle
\item SAA = list of sun azimuth angle
\item VZA = list of viewer zenith angle
\item VAA = list of viewer azimuth angle
}}
\usage{
get_s2_geometry_from_SAFE(s2xml)
}
\arguments{
\item{s2xml}{list. produced from reading XML metadata file with package XML}
}
\value{
List of S2 angles (SZA, SAA, VZA, VAA)
}
\description{
This function returns geometry of acquisition for S2 image processed with
Sen2Cor
\itemize{
\item SZA = list of sun zenith angle
\item SAA = list of sun azimuth angle
\item VZA = list of viewer zenith angle
\item VAA = list of viewer azimuth angle
}
}
