% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SENSOR.R
\name{get_spec_sensor}
\alias{get_spec_sensor}
\title{Computes spectral sensor characteristics for optical constants required for PROSPECT
soil parameters as well as direct and diffuse radiation for clear conditions}
\usage{
get_spec_sensor(
  SensorName = "MyCustomSensor",
  SpectralProps = NULL,
  Path_SensorResponse = NULL
)
}
\arguments{
\item{SensorName}{character. name of the sensor. Either already defined, or to be defined based on other inputs}

\item{SpectralProps}{list. Sensor spectral characteristics including wl = central wavelength and fwhm = corresponding FWHM}

\item{Path_SensorResponse}{character. Path where to get or save spectral response function}
}
\value{
SRF list. Spectral response function, corresponding spectral bands, and Original Bands
}
\description{
Computes spectral sensor characteristics for optical constants required for PROSPECT
soil parameters as well as direct and diffuse radiation for clear conditions
}
