% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_srf.R
\name{get_srf}
\alias{get_srf}
\title{Computes spectral response function based on wavelength and fwhm
characteristics}
\usage{
get_srf(wvl, fwhm, sensor_name = "user_defined")
}
\arguments{
\item{wvl}{numeric. spectral sampling of the sensor}

\item{fwhm}{numeric. Full Width Half Maximum for each spectral band}

\item{sensor_name}{character. sensor name}
}
\value{
SRF list. spectral response, sensor Spectral Bands & Original Bands
for which SRF is defined
}
\description{
Computes spectral response function based on wavelength and fwhm
characteristics
}
