% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_srf_sensor.R
\name{get_srf_sensor}
\alias{get_srf_sensor}
\title{reads spectral response from known sensor
spectral response from Sentinel-2 is already defined}
\usage{
get_srf_sensor(
  sensor_name = "user_defined",
  wl = NULL,
  fwhm = NULL,
  srf_path = "./",
  save_srf = TRUE
)
}
\arguments{
\item{sensor_name}{character. name of the sensor}

\item{wl}{numeric. central wavelength of spectral bands}

\item{fwhm}{numeric. fwhm of spectral bands defined by wl}

\item{srf_path}{character. path for the file where srf should be saved as CSV}

\item{save_srf}{boolean. Should srf be saved in file?}
}
\value{
srf list. Spectral response function, corresponding spectral bands,
and Original Bands
}
\description{
reads spectral response from known sensor
spectral response from Sentinel-2 is already defined
}
