% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_SENSOR.R
\name{GetRadiometry}
\alias{GetRadiometry}
\title{reads spectral response from known sensor
spectral response from Sentinel-2 is already defined}
\usage{
GetRadiometry(
  SensorName = "Custom",
  SpectralProps = NULL,
  Path_SensorResponse = "./",
  SaveSRF = TRUE
)
}
\arguments{
\item{SensorName}{character. name of the sensor}

\item{SpectralProps}{list. list of spectral properties including wl = central wavelength and fwhm = corresponding FWHM}

\item{Path_SensorResponse}{character. path for the file where SRF should be saved as CSV}

\item{SaveSRF}{boolean. Should SRF be saved in file?}
}
\value{
SRF list. Spectral response function, corresponding spectral bands, and Original Bands
}
\description{
reads spectral response from known sensor
spectral response from Sentinel-2 is already defined
}
