% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL_Inversion.R
\name{Invert_PROSAIL}
\alias{Invert_PROSAIL}
\title{Performs PROSAIL inversion using iterative optimization
in order to define estimate a set of user defined parameters}
\usage{
Invert_PROSAIL(brfMES, InitialGuess, LowerBound, UpperBound,
  SpecPROSPECT_Sensor, SpecATM_Sensor, SpecSOIL_Sensor, TypeLidf, ParmSet,
  MeritFunction = "Merit_RMSE_PROSAIL", PriorInfoMean = NULL,
  PriorInfoSD = NULL, WeightPrior = 0.01)
}
\arguments{
\item{brfMES}{numeric. measured BRF to be used during inversion}

\item{InitialGuess}{list. Initial guess of Parms2Estimate for the inversion}

\item{LowerBound}{list. Lower bound}

\item{UpperBound}{list. Upper bound}

\item{SpecPROSPECT_Sensor}{list. Includes optical constants
refractive index, specific absorption coefficients and corresponding spectral bands}

\item{SpecATM_Sensor}{list. direct and diffuse radiation for clear conditions}

\item{SpecSOIL_Sensor}{list. includes either dry soil and wet soil, or a unique soil sample if the psoil parameter is not inverted}

\item{TypeLidf}{numeric. Type of leaf inclination distribution function}

\item{ParmSet}{list. Parameters set to a fixed value by user}

\item{MeritFunction}{character. name of the function to be used as merit function
with given criterion to minimize (default = RMSE)}

\item{PriorInfoMean}{list. prior mean value of parameters defined as xprior}

\item{PriorInfoSD}{list. prior standard deviation of parameters defined as xprior}

\item{WeightPrior}{numeric. Weight to be applied on prior information to modulate its importance}
}
\value{
OutPROSPECT estimated values corresponding to Parms2Estimate
}
\description{
Performs PROSAIL inversion using iterative optimization
in order to define estimate a set of user defined parameters
}
