% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_PROSAIL.R
\name{invert_PROSAIL}
\alias{invert_PROSAIL}
\title{Performs PROSAIL inversion using iterative optimization
in order to define estimate a set of user defined parameters}
\usage{
invert_PROSAIL(
  brf_mes,
  initialization = NULL,
  lower_bound,
  upper_bound,
  SpecPROSPECT_Sensor,
  SpecATM_Sensor,
  SpecSOIL_Sensor,
  TypeLidf,
  parm_set,
  merit_function = "merit_RMSE_PROSAIL",
  prior_info = NULL
)
}
\arguments{
\item{brf_mes}{numeric. measured BRF to be used during inversion}

\item{initialization}{list. Initial guess of Parms2Estimate for the inversion}

\item{lower_bound}{list. Lower bound}

\item{upper_bound}{list. Upper bound}

\item{SpecPROSPECT_Sensor}{list. Includes optical constants
refractive index, specific absorption coefs and corresponding spectral bands}

\item{SpecATM_Sensor}{list. direct and diffuse radiation for clear conditions}

\item{SpecSOIL_Sensor}{list. includes either dry soil and wet soil,
or a unique soil sample if the psoil parameter is not inverted}

\item{TypeLidf}{numeric. Type of leaf inclination distribution function}

\item{parm_set}{list. Parameters set to a fixed value by user}

\item{merit_function}{character. name of the merit function
with given criterion to minimize (default = RMSE)}

\item{prior_info}{list. prior mean, sd and weight of parameters defined as xprior
modulate its importance}
}
\value{
OutPROSPECT estimated values corresponding to Parms2Estimate
}
\description{
Performs PROSAIL inversion using iterative optimization
in order to define estimate a set of user defined parameters
}
