% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfunc2.R
\name{jfunc2}
\alias{jfunc2}
\title{J2 function with avoidance of singularity problem}
\usage{
jfunc2(k, l, t)
}
\arguments{
\item{k}{numeric. Extinction coefficient for direct (solar or observer) flux}

\item{l}{numeric.}

\item{t}{numeric. Leaf Area Index}
}
\value{
Jout numeric.
}
\description{
J2 function with avoidance of singularity problem
}
