% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jfunc3.R
\name{Jfunc3}
\alias{Jfunc3}
\title{J3 function with avoidance of singularity problem}
\usage{
Jfunc3(k, l, t)
}
\arguments{
\item{k}{numeric. Extinction coefficient for direct (solar or observer) flux}

\item{l}{numeric.}

\item{t}{numeric. Leaf Area Index}
}
\value{
Jout numeric.
}
\description{
J3 function with avoidance of singularity problem
}
