% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL.R
\name{Jfunc4}
\alias{Jfunc4}
\title{J4 function for treating (near) conservative scattering}
\usage{
Jfunc4(m, t)
}
\arguments{
\item{m}{numeric. Extinction coefficient for direct (solar or observer) flux}

\item{t}{numeric. Leaf Area Index}
}
\value{
Jout numeric.
}
\description{
J4 function for treating (near) conservative scattering
}
