% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merit_RMSE_PROSAIL.R
\name{merit_rmse_prosail}
\alias{merit_rmse_prosail}
\title{Merit function for PROSAIL inversion}
\usage{
merit_rmse_prosail(
  xinit,
  parms_xinit,
  brf_mes,
  spec_prospect_sensor,
  spec_soil_sensor,
  spec_atm_sensor,
  parms_to_estimate,
  input_prosail,
  type_lidf,
  prior_info = NULL,
  parms_to_prior = NULL
)
}
\arguments{
\item{xinit}{numeric. Vector of input variables to estimate}

\item{parms_xinit}{character. name of parameters corresponding to xinit}

\item{brf_mes}{numeric. measured BRF}

\item{spec_prospect_sensor}{list. Includes optical constants for PROSPECT}

\item{spec_soil_sensor}{list. Includes soil reflectance (either 2 references
for optimization of psoil, or a unique spectrun)}

\item{spec_atm_sensor}{list. Includes diffuse and direct light
refractive index, specific absorption coefs and corresponding spectral bands}

\item{parms_to_estimate}{numeric. rank of variables to be inverted
to be estimated through inversion}

\item{input_prosail}{dataframe. full set of PROSAIL input variables}

\item{type_lidf}{numeric. type of leaf inclination distribution function}

\item{prior_info}{list. prior mean, sd & weight of parms defined as xprior}

\item{parms_to_prior}{numeric. rank of parameters to be used with prior info
to modulate its importance}
}
\value{
fc estimates of the parameters
}
\description{
Merit function for PROSAIL inversion
}
