% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL_Inversion.R
\name{Merit_RMSE_PROSAIL}
\alias{Merit_RMSE_PROSAIL}
\title{Merit function for PROSAIL inversion}
\usage{
Merit_RMSE_PROSAIL(xinit, brfMES, SpecPROSPECT_Sensor, SpecSOIL_Sensor,
  SpecATM_Sensor, Parms2Estimate, Parm2Set, ParmSet, InVar, TypeLidf)
}
\arguments{
\item{xinit}{numeric. Vector of input variables to estimate}

\item{brfMES}{numeric. measured BRF}

\item{SpecPROSPECT_Sensor}{list. Includes optical constants for PROSPECT}

\item{SpecSOIL_Sensor}{list. Includes soil reflectance (either 2 references for optimization of psoil, or a unique spectrun)}

\item{SpecATM_Sensor}{list. Includes diffuse and direct light
refractive index, specific absorption coefficients and corresponding spectral bands}

\item{Parms2Estimate}{numeric. rank of variables to be inverted
to be estimated through inversion}

\item{Parm2Set}{numeric. rank of variables to be set out of inversion}

\item{ParmSet}{list. value of variables to be set out of inversion}

\item{InVar}{dataframe. full set of PROSAIL input variables}

\item{TypeLidf}{numeric. tyoe of leaf inclination distribution function}
}
\value{
fc estimates of the parameters
}
\description{
Merit function for PROSAIL inversion
}
