% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sensor_simulation.R
\name{prepare_sensor_simulation}
\alias{prepare_sensor_simulation}
\title{This function converts high resolution spectral info into broader
spectral characteristics}
\usage{
prepare_sensor_simulation(SpecPROSPECT, SpecSOIL, SpecATM, SRF)
}
\arguments{
\item{SpecPROSPECT}{list. optical constants required for PROSPECT}

\item{SpecSOIL}{list. dry soil and wet soil, or a unique soil sample if
psoil parameter is not inverted}

\item{SpecATM}{list. direct and diffuse radiation for clear conditions}

\item{SRF}{list. Spectral response function, corresponding spectral bands,
and Original Bands}
}
\value{
SpecSensor list. list of input specs with sensor resolution
}
\description{
This function converts high resolution spectral info into broader
spectral characteristics
}
