% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prosail.R
\name{prosail}
\alias{prosail}
\title{PROSAIL simulation based on a set of combinations of input parameters}
\usage{
prosail(
  spec_sensor = NULL,
  input_prospect = NULL,
  n_struct = 1.5,
  chl = 40,
  car = 8,
  ant = 0,
  brown = 0,
  ewt = 0.01,
  lma = NULL,
  prot = 0,
  cbc = 0,
  alpha = 40,
  type_lidf = 2,
  lidf_a = 60,
  lidf_b = NULL,
  lai = 3,
  q = 0.1,
  tts = 30,
  tto = 0,
  psi = 60,
  rsoil = NULL,
  fraction_brown = 0,
  diss = 0,
  cv = 1,
  zeta = 1,
  SAILversion = "4SAIL",
  brown_lop = NULL
)
}
\arguments{
\item{spec_sensor}{list. Includes optical constants required for PROSPECT
refractive index, specific absorption coefficients and spectral bands}

\item{input_prospect}{list. PROSPECT input variables}

\item{n_struct}{numeric. Leaf structure parameter}

\item{chl}{numeric. chlorophyll content (microg.cm-2)}

\item{car}{numeric. carotenoid content (microg.cm-2)}

\item{ant}{numeric. anthocyain content (microg.cm-2)}

\item{brown}{numeric. brown pigment content (Arbitrary units)}

\item{ewt}{numeric. Equivalent Water Thickness (g.cm-2)}

\item{lma}{numeric. Leaf Mass per Area (g.cm-2)}

\item{prot}{numeric. protein content  (g.cm-2)}

\item{cbc}{numeric. NonProtCarbon-based constituent content (g.cm-2)}

\item{alpha}{numeric. Solid angle for incident light at surface of leaf}

\item{type_lidf}{numeric. Type of leaf inclination distribution function}

\item{lidf_a}{numeric.
if type_lidf ==1, controls the average leaf slope
if type_lidf ==2, controls the average leaf angle}

\item{lidf_b}{numeric.
if type_lidf ==1, controls the distribution's bimodality
if type_lidf ==2, unused}

\item{lai}{numeric. Leaf Area Index}

\item{q}{numeric. Hot Spot parameter}

\item{tts}{numeric. Sun zeith angle}

\item{tto}{numeric. Observer zeith angle}

\item{psi}{numeric. Azimuth Sun / Observer}

\item{rsoil}{numeric. Soil reflectance}

\item{fraction_brown}{numeric. Fraction of brown leaf area}

\item{diss}{numeric. Layer dissociation factor}

\item{cv}{numeric. vertical crown cover percentage
= \% ground area covered with crowns as seen from nadir direction}

\item{zeta}{numeric. Tree shape factor
= ratio of crown diameter to crown height}

\item{SAILversion}{character. choose between 4SAIL and 4SAIL2}

\item{brown_lop}{dataframe. optical properties for brown vegetation
\itemize{
\item WVL, Reflectance, Transmittance
\item Set to NULL if use PROSPECT to generate it
}}
}
\value{
list. rdot,rsot,rddt,rsdt
rdot: hemispherical-directional reflectance factor in viewing direction
rsot: bi-directional reflectance factor
rsdt: directional-hemispherical reflectance factor for solar incident flux
rddt: bi-hemispherical reflectance factor
}
\description{
PROSAIL simulation based on a set of combinations of input parameters
}
