% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prosail_hybrid_apply.R
\name{prosail_hybrid_apply}
\alias{prosail_hybrid_apply}
\title{This function applies the regression models trained with prosail_hybrid_train}
\usage{
prosail_hybrid_apply(regression_models, refl, progressBar = FALSE)
}
\arguments{
\item{regression_models}{list. regression models produced by
prosail_hybrid_train}

\item{refl}{numeric. LUT of BRF used for training}

\item{progressBar}{boolean. should progressbar be displayed?}
}
\value{
HybridRes list. Estimated values corresponding to refl. Includes
\itemize{
\item MeanEstimate = mean value for the ensemble regression model
\item StdEstimate = std value for the ensemble regression model
}
}
\description{
This function applies the regression models trained with prosail_hybrid_train
}
