% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROSAIL_Hybrid_Train.R
\name{prosail_hybrid_train}
\alias{prosail_hybrid_train}
\title{This function trains a regression model to estimate a set of variables
from spectral data}
\usage{
prosail_hybrid_train(
  brf_lut,
  input_variables,
  nb_bagg = 20,
  replacement = FALSE,
  method = "liquidSVM",
  verbose = FALSE,
  progressBar = FALSE,
  options = NULL
)
}
\arguments{
\item{brf_lut}{numeric. LUT of BRF used for training}

\item{input_variables}{numeric. biophysical parameter corresponding to refl}

\item{nb_bagg}{numeric. nb of individual subsets generated from brf_lut}

\item{replacement}{Boolean. subsets generated with / without replacement}

\item{method}{character. which machine learning regression method used?
default = SVM with liquidSVM.
set nu-svr for RBF SVM with kernlab package
set svmRadial or svmLinear from SVM with caret package
More to come}

\item{verbose}{boolean. when set to TRUE, prints message if hyperparameter
adjustment performed during training}

\item{progressBar}{boolean. should progressbar be displayed?}

\item{options}{list}
}
\value{
models_mlr list. ML regression models trained for the retrieval of
input_variables based on brf_lut
}
\description{
This function trains a regression model to estimate a set of variables
from spectral data
}
