% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROSAIL_Hybrid_Train.R
\name{PROSAIL_Hybrid_Train}
\alias{PROSAIL_Hybrid_Train}
\title{This function trains a support vector regression for a set of variables based on spectral data}
\usage{
PROSAIL_Hybrid_Train(
  BRF_LUT,
  InputVar,
  nbEnsemble = 20,
  WithReplacement = FALSE,
  method = "liquidSVM",
  verbose = FALSE,
  progressBar = FALSE
)
}
\arguments{
\item{BRF_LUT}{numeric. LUT of bidirectional reflectances factors used for training}

\item{InputVar}{numeric. biophysical parameter corresponding to the reflectance}

\item{nbEnsemble}{numeric. Number of individual subsets should be generated from BRF_LUT}

\item{WithReplacement}{Boolean. should subsets be generated with or without replacement?}

\item{method}{character. which machine learning regression method should be used?
default = SVM with liquidSVM. svmRadial and svmLinear from caret package also implemented. More to come}

\item{verbose}{boolean. when set to TRUE, prints message if hyperparameter adjustment performed during training}

\item{progressBar}{boolean. should progressbar be displayed?}
}
\value{
modelsMLR list. ML regression models trained for the retrieval of
InputVar based on BRF_LUT
}
\description{
This function trains a support vector regression for a set of variables based on spectral data
}
