% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_compatibility_input_prosail.R
\name{set_compatibility_input_prosail}
\alias{set_compatibility_input_prosail}
\title{make sure of the compatibility of variable names between versions of prosail}
\usage{
set_compatibility_input_prosail(input_prosail)
}
\arguments{
\item{input_prosail}{list. PROSAIL input variables}
}
\value{
input_prosail numeric. updated input_prosail
}
\description{
make sure of the compatibility of variable names between versions of prosail
}
