% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRF_availability.R
\name{srf_availability}
\alias{srf_availability}
\title{This function returns the list of al sensors for which SRF is available}
\usage{
srf_availability()
}
\value{
list of sensors
}
\description{
This function returns the list of al sensors for which SRF is available
}
