% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lib_PROSAIL_HybridInversion.R
\name{train_prosail_inversion}
\alias{train_prosail_inversion}
\title{This function performs full training for hybrid invrsion using SVR with
values for default parameters}
\usage{
train_prosail_inversion(
  minval = NULL,
  maxval = NULL,
  TypeDistrib = NULL,
  GaussianDistrib = NULL,
  ParmSet = NULL,
  nbSamples = 2000,
  nbSamplesPerRun = 100,
  nbModels = 20,
  Replacement = TRUE,
  SAILversion = "4SAIL",
  Parms2Estimate = "lai",
  Bands2Select = NULL,
  NoiseLevel = NULL,
  SpecPROSPECT = NULL,
  SpecSOIL = NULL,
  SpecATM = NULL,
  Path_Results = "./",
  FigPlot = FALSE
)
}
\arguments{
\item{minval}{list. minimum value for input parameters sampled to produce a training LUT}

\item{maxval}{list. maximum value for input parameters sampled to produce a training LUT}

\item{TypeDistrib}{list. Type of distribution. Either 'Uniform' or 'Gaussian'}

\item{GaussianDistrib}{list. Mean value and STD corresponding to the parameters sampled with gaussian distribution}

\item{ParmSet}{list. list of input parameters set to a specific value}

\item{nbSamples}{numeric. number of samples in training LUT}

\item{nbSamplesPerRun}{numeric. number of training sample per individual regression model}

\item{nbModels}{numeric. number of individual models to be run for ensemble}

\item{Replacement}{bolean. is there replacement in subsampling?}

\item{SAILversion}{character. Either 4SAIL or 4SAIL2}

\item{Parms2Estimate}{list. list of input parameters to be estimated}

\item{Bands2Select}{list. list of bands used for regression for each input parameter}

\item{NoiseLevel}{list. list of noise value added to reflectance (defined per input parm)}

\item{SpecPROSPECT}{list. Includes optical constants required for PROSPECT}

\item{SpecSOIL}{list. Includes either dry soil and wet soil, or a unique soil sample if the psoil parameter is not inverted}

\item{SpecATM}{list. Includes direct and diffuse radiation for clear conditions}

\item{Path_Results}{character. path for results}

\item{FigPlot}{boolean. Set TRUE to get scatterplot of estimated biophysical variable during training step}
}
\value{
modelsSVR list. regression models trained for the retrieval of InputVar based on BRF_LUT
}
\description{
This function performs full training for hybrid invrsion using SVR with
values for default parameters
}
