% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_prosail_inversion.R
\name{train_prosail_inversion}
\alias{train_prosail_inversion}
\title{This function performs full training for hybrid inversion using SVR with
values for default parameters}
\usage{
train_prosail_inversion(
  input_prosail = NULL,
  BRF_LUT = NULL,
  atbd = FALSE,
  GeomAcq = NULL,
  SRF = NULL,
  codistribution_lai = TRUE,
  minval = NULL,
  maxval = NULL,
  TypeDistrib = NULL,
  GaussianDistrib = NULL,
  parm_set = NULL,
  SAILversion = "4SAIL",
  brown_lop = NULL,
  nb_samples = 2000,
  nb_models = 20,
  replacement = TRUE,
  Parms2Estimate = "lai",
  selected_bands = NULL,
  noise_level = NULL,
  SpecPROSPECT = NULL,
  SpecSOIL = NULL,
  SpecATM = NULL,
  output_dir = "./",
  method = "liquidSVM",
  verbose = FALSE
)
}
\arguments{
\item{input_prosail}{list. user-defined list of input parameters to be used
to produce a training LUT}

\item{BRF_LUT}{list. user-defined BRF LUT used to run the hybrid inversion}

\item{atbd}{boolean. should input parameter distribution from ATBD be
applied ?}

\item{GeomAcq}{list. geometry of acquisition: min and max tts, tto & psi}

\item{SRF}{list. Spectral response function}

\item{codistribution_lai}{boolean. set TRUE if codistribution with LAI accounted for}

\item{minval}{list. min val for input parms sampled to produce a training LUT}

\item{maxval}{list. max val for input parms sampled to produce a training LUT}

\item{TypeDistrib}{list. Type of distribution. 'Uniform' or 'Gaussian'}

\item{GaussianDistrib}{list. Mean value and STD corresponding to the
parameters sampled with gaussian distribution}

\item{parm_set}{list. list of input parameters set to a specific value}

\item{SAILversion}{character. Either 4SAIL or 4SAIL2}

\item{brown_lop}{character. Either 4SAIL or 4SAIL2}

\item{nb_samples}{numeric. number of samples in training LUT}

\item{nb_models}{numeric. number of individual models to be run for ensemble}

\item{replacement}{bolean. is there replacement in subsampling?}

\item{Parms2Estimate}{list. list of input parameters to be estimated}

\item{selected_bands}{list. bands used for regression for each input parameter}

\item{noise_level}{list. noise added to reflectance (defined per input parm)}

\item{SpecPROSPECT}{list. Includes optical constants required for PROSPECT}

\item{SpecSOIL}{list. Includes either dry soil and wet soil, or a unique
soil sample if the psoil parameter is not inverted}

\item{SpecATM}{list. direct and diffuse radiation for clear conditions}

\item{output_dir}{character. path for results
variable during training step}

\item{method}{character. which machine learning regression method should be
used?
default = SVM with liquidSVM. svmRadial and svmLinear from caret package also
implemented. More to come}

\item{verbose}{boolean. when set to TRUE, prints message if hyperparameter
adjustment performed during training}
}
\value{
modelsSVR list. regression models trained for the retrieval of
input_variables based on BRF_LUT
}
\description{
This function performs full training for hybrid inversion using SVR with
values for default parameters
}
