% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_parm_prior.R
\name{which_parm_prior}
\alias{which_parm_prior}
\title{function identifying for which parameters prior information is known}
\usage{
which_parm_prior(PriorInfoMean, PriorInfoSD)
}
\arguments{
\item{PriorInfoMean}{list. mean value expected for a list of parameters}

\item{PriorInfoSD}{list. dtandard deviation to the mean expected for a list
of parameters}
}
\value{
fc estimates of the parameters
}
\description{
function identifying for which parameters prior information is known
}
