% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_envi_header.R
\name{write_envi_header}
\alias{write_envi_header}
\title{writes ENVI hdr file}
\usage{
write_envi_header(hdr, hdr_path)
}
\arguments{
\item{hdr}{content to be written}

\item{hdr_path}{Path of the hdr file}
}
\value{
None
}
\description{
writes ENVI hdr file
}
